### 2025-05-21 #8256 @editorialbot
[PRE REVIEW]: DDC: The Discrete Domain Computation library
**Submitting author:** @tpadioleau (<a href="http://orcid.org/0000-0001-5496-0013">Thomas Padioleau</a>)
**Repository:** https://github.com/CExA-project/ddc
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.7.0
