+++ search/functions_0.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['accessor_0',['accessor',['../classddc_1_1ChunkCommon.html#aa8b0804738bd20206cea06c2f91177e5',1,'ddc::ChunkCommon::accessor()'],['../classddc_1_1ChunkCommon_3_01ElementType_00_01DiscreteDomain_3_01DDims_8_8_8_01_4_00_01LayoutStridedPolicy_01_4.html#abd9a059224600a1f395b0aabb4912829',1,'ddc::ChunkCommon&lt; ElementType, DiscreteDomain&lt; DDims... &gt;, LayoutStridedPolicy &gt;::accessor()']]],
+  ['accessor_0',['accessor',['../classddc_1_1ChunkCommon.html#aa8b0804738bd20206cea06c2f91177e5',1,'ddc::ChunkCommon']]],
   ['alignedallocator_1',['alignedallocator',['../classddc_1_1AlignedAllocator.html#ac230d56372c1220184f774e6e0a1ca2a',1,'ddc::AlignedAllocator::AlignedAllocator()=default'],['../classddc_1_1AlignedAllocator.html#a7ee3cba307070f81ccd8d9e245a668a3',1,'ddc::AlignedAllocator::AlignedAllocator(AlignedAllocator const &amp;x)=default'],['../classddc_1_1AlignedAllocator.html#ab494004552fdaee4fd5c5e0cabd9c1db',1,'ddc::AlignedAllocator::AlignedAllocator(AlignedAllocator &amp;&amp;x) noexcept=default'],['../classddc_1_1AlignedAllocator.html#a0128fefb3baa896fc60ead781bdcd941',1,'ddc::AlignedAllocator::AlignedAllocator(AlignedAllocator&lt; U, N &gt; const &amp;) noexcept']]],
   ['allocate_2',['allocate',['../classddc_1_1AlignedAllocator.html#aeb340bbe2aac27ca233dd3f3a2ddbd49',1,'ddc::AlignedAllocator::allocate()'],['../classddc_1_1KokkosAllocator.html#a2e365ed5fc56fdbc3ccc5ae3f96a5ad8',1,'ddc::KokkosAllocator::allocate(std::size_t n) const'],['../classddc_1_1KokkosAllocator.html#af266592d558928dff4a351f2b6a77c7a',1,'ddc::KokkosAllocator::allocate(std::string const &amp;label, std::size_t n) const']]],
[…]
