+++ search/typedefs_4.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['element_5ftype_0',['element_type',['../classddc_1_1Chunk.html#a985a822d0ce798211adabdebc7b81270',1,'ddc::Chunk::element_type'],['../classddc_1_1ChunkCommon_3_01ElementType_00_01DiscreteDomain_3_01DDims_8_8_8_01_4_00_01LayoutStridedPolicy_01_4.html#ade37a22ad54c85d4a0b819fa3d6b3d36',1,'ddc::ChunkCommon&lt; ElementType, DiscreteDomain&lt; DDims... &gt;, LayoutStridedPolicy &gt;::element_type'],['../classddc_1_1ChunkCommon.html#ae05e9684816f81523f10da427f5735fd',1,'ddc::ChunkCommon::element_type'],['../classddc_1_1ChunkSpan_3_01ElementType_00_01DiscreteDomain_3_01DDims_8_8_8_01_4_00_01LayoutStridedPolicy_00_01MemorySpace_01_4.html#aecb8343330039f5ababe28630b08b03c',1,'ddc::ChunkSpan&lt; ElementType, DiscreteDomain&lt; DDims... &gt;, LayoutStridedPolicy, MemorySpace &gt;::element_type'],['../classddc_1_1ChunkSpan.html#ade4fbbb2fe1b7a816fc2641d6d837fcc',1,'ddc::ChunkSpan::element_type'],['../classddc_1_1Chunk_3_01ElementType_00_01DiscreteDomain_3_01DDims_8_8_8_01_4_00_01Allocator_01_4.html#a0a9c9a52b66d51d0d6bbd278e97206c6',1,'ddc::Chunk&lt; ElementType, DiscreteDomain&lt; DDims... &gt;, Allocator &gt;::element_type']]],
+  ['element_5ftype_0',['element_type',['../classddc_1_1ChunkCommon.html#ae05e9684816f81523f10da427f5735fd',1,'ddc::ChunkCommon::element_type'],['../classddc_1_1ChunkSpan.html#ade4fbbb2fe1b7a816fc2641d6d837fcc',1,'ddc::ChunkSpan::element_type'],['../classddc_1_1Chunk.html#a985a822d0ce798211adabdebc7b81270',1,'ddc::Chunk::element_type']]],
   ['evaluation_5fdiscrete_5fdimension_5ftype_1',['evaluation_discrete_dimension_type',['../classddc_1_1SplineEvaluator.html#aaff5facf824889023832086180900f05',1,'ddc::SplineEvaluator']]],
   ['evaluation_5fdiscrete_5fdimension_5ftype1_2',['evaluation_discrete_dimension_type1',['../classddc_1_1SplineEvaluator2D.html#ac97aa4ee0f9472ff2911b885a20f9ac4',1,'ddc::SplineEvaluator2D']]],
   ['evaluation_5fdiscrete_5fdimension_5ftype2_3',['evaluation_discrete_dimension_type2',['../classddc_1_1SplineEvaluator2D.html#aaef2585bd9a33e8f7bc5c77111ce01f6',1,'ddc::SplineEvaluator2D']]],
@@ -8,5 +8,5 @@ var searchData=
   ['evaluation_5fdomain_5ftype1_5',['evaluation_domain_type1',['../classddc_1_1SplineEvaluator2D.html#a1dd5fb0e5efdc1774d85446f6ea100a8',1,'ddc::SplineEvaluator2D']]],
[…]

[…]
