### 2025-05-21 #8256 @tpadioleau
Thank you for your remarks.

> 👋 [@tpadioleau](https://github.com/tpadioleau) - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

It seems we have a conflict between reuse tools and the tool building the pdf. We will try to find a way to make them work together.

> In addition, your paper is almost twice as long as the maximum recommended JOSS paper length of 1000 words. You will need to reduce the paper, perhaps by pointing at material in the documentation or repo rather than including it in the paper.

We will try to reduce it further.

> I also don't understand why you have two licenses, only one of which is discussed in the copyright file

DDC was first released under MIT license. When we added the file `cmake/FindLAPACKE.cmake` licensed under CECILL-C, following the REUSE tools in the CI, this introduced a new license in our repository. I think we forgot to update the file `COPYRIGHT.md`. I wonder if we should we just remove this file ? Remove the name of licenses in this file ? We would gladly take your recommendations if you have any.
