### 2025-05-21 #8256 @danielskatz
> DDC was first released under MIT license. When we added the file `cmake/FindLAPACKE.cmake` licensed under CECILL-C, following the REUSE tools in the CI, this introduced a new license in our repository. I think we forgot to update the file `COPYRIGHT.md`. I wonder if we should we just remove this file ? Remove the name of licenses in this file ? We would gladly take your recommendations if you have any.

I'm not an expert on licenses, but this might be as simple as having a file that explains what licenses apply to what parts of the software.

Because both licenses are OSI-approved, JOSS can review the submission.
