### 2025-05-21 #8256 @danielskatz
> It seems we have a conflict between reuse tools and the tool building the pdf. We will try to find a way to make them work together.

I think the probably might be simpler, just that your header doesn't follow the example of the JOSS header. For example, you have a bunch of blank lines, and you also have some comments at the start (which are probably ok). I would first just remove the blank lines in the metadata section.
