### 2025-05-22 539367d
Move licensing out of the file
