+++ paper/paper.bib
@@ -117,20 +117,6 @@ url = {https://github.com/kokkos/kokkos-fft},
 version = {0.3.0}
 }
 
-@software{pdi,
-author = {{The PDI team}},
-title = {{PDI}},
-url = {https://github.com/pdidev/pdi},
-version = {1.9.1}
-}
-
-@software{xrft,
-author = {{The xrft team}},
-title = {{xrft}},
-url = {https://github.com/xgcm/xrft},
-version = {1.0.1}
-}
-
 @software{Pusz_mp-units_2024,
 author = {Pusz, Mateusz and Guerrero Peña, Johel Ernesto and Hogg, Chip and {The mp-units project team}},
 license = {MIT},
