+++ paper/paper.md
@@ -55,13 +55,6 @@ The library is actively used to modernize the Fortran-based Gysela plasma simula
 
 The DDC library is a C++ library designed for expressive and safe handling of multidimensional data. Its core component provides flexible data containers along with algorithms built on top of the performance portable Kokkos library. The library is fully compatible with Kokkos and does not attempt to hide it, allowing users to leverage Kokkos' full capabilities while benefiting from DDC’s strongly-typed, labelled multidimensional arrays when and where it makes sense.
 
-### Containers
-
-DDC offers two multidimensional containers designed over the C++ 23 multidimensional array `std::mdspan`:
-
-- `Chunk` an owning container, i.e. it manages the lifetime of the underlying memory allocation,
-- `ChunkSpan` is a non-owning container view on an existing memory allocation.
-
 ### Strongly-typed labelled indices
 
 DDC employs strongly-typed multi-indices to label dimensions and access data. It introduces two types of multi-indices to access the container's data:
@@ -98,113 +91,11 @@ This highlights the fact that `DiscreteElement` provides a globally consistent i
 
 The semantics of DDC containers associates data to a set of `DiscreteElement` indices. Let us note that the set of all possible `DiscreteElement` has a total order that is typically established once and for all at program initialization. Thus to be able to construct a DDC container one must provide a multidimensional set of `DiscreteElement` indices, only these indices can be later used to access the container’s data.
 
-The library provides multiple ways to group `DiscreteElement`s into sets. Each set takes the form of a Cartesian product of the sets along each dimensions. DDC predefines the following sets:
-
-- `DiscreteDomain`, a Cartesian product of intervals of `DiscreteElement` in each dimension,
-- `StridedDiscreteDomain`, a Cartesian product of sets of `DiscreteElement` with a fixed step/stride in each dimension,
-- `SparseDiscreteDomain`, a Cartesian product of explicitly ordered lists of `DiscreteElement` in each dimension.
-
-The performance of the container's data access depends on the properties of the set considered. Indeed the set is used to retrieve the position of a given multi-index relative to the front multi-index. Thus the performance of this operation depends on the information available in the set.
-
-### Slicing
-
-Similar to `std::mdspan`, DDC containers support slicing through the bracket operator that always returns a `ChunkSpan`. The key property is that the resulting slice’s set of `DiscreteElement` is a subset of the original set.
+The library provides several ways to group `DiscreteElement` into sets, each represented as a Cartesian product of per-dimension sets. These sets offer a lookup function to retrieve the position of a multi-index relative to the front of the set. The performance of container data access depends significantly on the compile-time properties of the set used.
 
 ### Multidimensional algorithms
 
-Finally, DDC offers multidimensional algorithms to manipulate the containers and associated `DiscreteElement` indices. The parallel versions are based on Kokkos providing performance portability. Here is a list as of version 0.7:
-
-- `parallel_deepcopy`, copies two `ChunkSpan` that are defined on the same set of `DiscreteElement`,
-- `parallel_for_each`, applies a function once per `DiscreteElement` of a set of `DiscreteElement`,
-- `parallel_fill`, fills a `ChunkSpan` with a value,
-- `parallel_transform_reduce`, applies a function once per `DiscreteElement` of a set of `DiscreteElement` and combines the returned values with a reducer.
-
-## Example
-
[…]
