### 2025-05-22 #8256 @tpadioleau
WIP to reduce the size
