### 2025-05-22 8149149
Remove unused examples to speed-up the CI
