### 2025-05-22 46cc424
Remove unused examples to speed-up the CI
