+++ paper/paper.bib
@@ -79,7 +79,7 @@ volume = {207},
 pages = {35-68},
 year = {2016},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2016.05.007},
+doi = {10.1016/j.cpc.2016.05.007},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465516301230},
 author = {V. Grandgirard and J. Abiteboul and J. Bigot and T. Cartier-Michaud and N. Crouseilles and G. Dif-Pradalier and Ch. Ehrlacher and D. Esteve and X. Garbet and Ph. Ghendrih and G. Latu and M. Mehrenberger and C. Norscini and Ch. Passeron and F. Rozar and Y. Sarazin and E. Sonnendrücker and A. Strugarek and D. Zarzoso},
 }
