### 2025-05-23 b12451e
Add an explanation about the multiple licenses (#859)
