+++ COPYRIGHT.md
@@ -9,7 +9,9 @@ SPDX-License-Identifier: MIT
 The DDC project licencing information follows the REUSE specification.
 
 It is distributed under the MIT License whose full text is available in the
-LICENSES/MIT.txt file.
+LICENSES/MIT.txt file. A small number of files in this repository do not
+originate from the DDC developers. These files may have their own licences
+which will be indicated in the headers.
 
 For the purpose of copyright, the DDC development team is defined as follow:
 
