### 2025-05-23 #8256 @tpadioleau
@danielskatz thank you for your comments. I think we have addressed all issues:
- the paper has been much reduced, down to 1275 words, while keeping the important key messages,
- invalid DOIs have been fixed,
- the paper can now be built from the editorialbot: we have put the licensing header out of the paper.

Please let us know if we need to do further modifications.
