### 2025-05-26 5154fdf
Rework assertions
