+++ tests/chunk.cpp
@@ -525,9 +525,11 @@ TEST(Chunk2DTest, SliceCoordXOutOfBounds)
 #if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
     ChunkXY<double> chunk(dom_x_y);
 
-    EXPECT_DEATH(
-            chunk[dom_x.back() + 1],
-            R"rgx([Aa]ssert.*select<QueryDDims...>\(this->m_domain\).contains\(slice_spec\))rgx");
+    char const* const death_msg
+            = R"rgx([Aa]ssert.*select<QueryDDims...>\(this->m_domain\).contains\(slice_spec\))rgx";
+
+    EXPECT_DEATH(chunk[dom_x.front() - 1], death_msg);
+    EXPECT_DEATH(chunk[dom_x.back() + 1], death_msg);
 #else
     GTEST_SKIP();
 #endif
@@ -588,11 +590,13 @@ TEST(Chunk2DTest, SliceDomainY)
 TEST(Chunk2DTest, SliceDomainXOutOfBounds)
 {
 #if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
-    ChunkXY<double> chunk(dom_x_y.remove_first(DVectXY(1, 0)));
+    ChunkXY<double> chunk(dom_x_y.remove(DVectXY(1, 0), DVectXY(1, 0)));
 
-    EXPECT_DEATH(
-            chunk[dom_x_y],
-            R"rgx([Aa]ssert.*DiscreteDomain<QueryDDims...>\(this->m_domain\).contains\(odomain.front\(\)\) && DiscreteDomain<QueryDDims...>\(this->m_domain\).contains\(odomain.back\(\)\))rgx");
+    char const* const death_msg
+            = R"rgx([Aa]ssert.*DiscreteDomain<QueryDDims...>\(this->m_domain\).contains\(odomain.front\(\)\) && DiscreteDomain<QueryDDims...>\(this->m_domain\).contains\(odomain.back\(\)\))rgx";
+
+    EXPECT_DEATH(chunk[dom_x_y.remove_last(DVectXY(1, 0))], death_msg);
+    EXPECT_DEATH(chunk[dom_x_y.remove_first(DVectXY(1, 0))], death_msg);
 #else
     GTEST_SKIP();
 #endif
