+++ include/ddc/ddc_to_kokkos_execution_policy.hpp
@@ -9,9 +9,10 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "ddc/detail/kokkos.hpp"
-#include "ddc/discrete_domain.hpp"
-#include "ddc/discrete_element.hpp"
+#include "detail/kokkos.hpp"
+
+#include "discrete_domain.hpp"
+#include "discrete_element.hpp"
 
 namespace ddc::detail {
 
