+++ include/ddc/discrete_element.hpp
@@ -12,9 +12,10 @@
 
 #include <Kokkos_Macros.hpp>
 
-#include "ddc/detail/macros.hpp"
-#include "ddc/detail/type_seq.hpp"
-#include "ddc/discrete_vector.hpp"
+#include "detail/macros.hpp"
+#include "detail/type_seq.hpp"
+
+#include "discrete_vector.hpp"
 
 namespace ddc {
 
