+++ include/ddc/parallel_transform_reduce.hpp
@@ -11,11 +11,12 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "ddc/ddc_to_kokkos_execution_policy.hpp"
-#include "ddc/detail/kokkos.hpp"
-#include "ddc/discrete_domain.hpp"
-#include "ddc/discrete_element.hpp"
-#include "ddc/reducer.hpp"
+#include "detail/kokkos.hpp"
+
+#include "ddc_to_kokkos_execution_policy.hpp"
+#include "discrete_domain.hpp"
+#include "discrete_element.hpp"
+#include "reducer.hpp"
 
 namespace ddc {
 
