+++ include/ddc/chunk.hpp
@@ -11,12 +11,13 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "ddc/chunk_common.hpp"
-#include "ddc/chunk_span.hpp"
-#include "ddc/chunk_traits.hpp"
-#include "ddc/detail/kokkos.hpp"
-#include "ddc/detail/type_traits.hpp"
-#include "ddc/kokkos_allocator.hpp"
+#include "detail/kokkos.hpp"
+#include "detail/type_traits.hpp"
+
+#include "chunk_common.hpp"
+#include "chunk_span.hpp"
+#include "chunk_traits.hpp"
+#include "kokkos_allocator.hpp"
 
 namespace ddc {
 
