+++ include/ddc/parallel_deepcopy.hpp
@@ -9,7 +9,7 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "ddc/chunk_traits.hpp"
+#include "chunk_traits.hpp"
 
 namespace ddc {
 
