### 2025-05-27 1277ca6
Add assertions when slicing (#860)
* Add assertions when slicing

* Rework assertions
