+++ tests/batch.cpp
@@ -1,0 +1,210 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include <iostream>
+
+#include <ddc/ddc.hpp>
+
+#include <gtest/gtest.h>
+
+#include <Kokkos_Core.hpp>
+
+inline namespace anonymous_namespace_workaround_mean_cpp {
+
+using DElem0D = ddc::DiscreteElement<>;
+using DVect0D = ddc::DiscreteVector<>;
+using DDom0D = ddc::DiscreteDomain<>;
+
+template <class Datatype>
+using Chunk0D = ddc::Chunk<Datatype, DDom0D>;
+template <class Datatype>
+using ChunkSpan0D = ddc::ChunkSpan<Datatype, DDom0D>;
+
+
+struct DDimX
+{
+};
+using DElemX = ddc::DiscreteElement<DDimX>;
+using DVectX = ddc::DiscreteVector<DDimX>;
+using DDomX = ddc::DiscreteDomain<DDimX>;
+
+template <class Datatype>
+using ChunkX = ddc::Chunk<Datatype, DDomX>;
+
+
+struct DDimY
+{
+};
+using DElemY = ddc::DiscreteElement<DDimY>;
+using DVectY = ddc::DiscreteVector<DDimY>;
+using DDomY = ddc::DiscreteDomain<DDimY>;
+
+template <class Datatype>
+using ChunkY = ddc::Chunk<Datatype, DDomY>;
+
+
+struct DDimZ
+{
+};
+using DElemZ = ddc::DiscreteElement<DDimZ>;
+using DVectZ = ddc::DiscreteVector<DDimZ>;
+using DDomZ = ddc::DiscreteDomain<DDimZ>;
+
+
+using DElemXY = ddc::DiscreteElement<DDimX, DDimY>;
+using DVectXY = ddc::DiscreteVector<DDimX, DDimY>;
+using DDomXY = ddc::DiscreteDomain<DDimX, DDimY>;
+
+template <class Datatype>
+using ChunkXY = ddc::Chunk<Datatype, DDomXY>;
+
+
+using DElemYX = ddc::DiscreteElement<DDimY, DDimX>;
+using DVectYX = ddc::DiscreteVector<DDimY, DDimX>;
+using DDomYX = ddc::DiscreteDomain<DDimY, DDimX>;
+
+template <class Datatype>
+using ChunkYX = ddc::Chunk<Datatype, DDomYX>;
+
+
+DElem0D constexpr lbound_0d {};
+DVect0D constexpr nelems_0d {};
+DDom0D constexpr dom_0d(lbound_0d, nelems_0d);
+
+DElemX constexpr lbound_x = ddc::init_trivial_half_bounded_space<DDimX>();
+DVectX constexpr nelems_x(3);
+DDomX constexpr dom_x(lbound_x, nelems_x);
+
+DElemY constexpr lbound_y = ddc::init_trivial_half_bounded_space<DDimY>();
+DVectY constexpr nelems_y(12);
+
+DElemZ constexpr lbound_z = ddc::init_trivial_half_bounded_space<DDimZ>();
+
+DElemXY constexpr lbound_x_y(lbound_x, lbound_y);
+DVectXY constexpr nelems_x_y(nelems_x, nelems_y);
+DDomXY constexpr dom_x_y(lbound_x_y, nelems_x_y);
+
+} // namespace anonymous_namespace_workaround_mean_cpp
+
+template <
+        typename Tin,
+        typename Tout,
+        typename DDomIn,
+        typename DDomOut,
+        typename ExecSpace,
+        typename MemorySpace,
+        typename LayoutIn,
[…]
