### 2025-05-28 ff550d7
Add parallel_copy algorithm
