### 2025-06-02 5566270
Bump mikepenz/action-junit-report from 5.5.1 to 5.6.0
Bumps [mikepenz/action-junit-report](https://github.com/mikepenz/action-junit-report) from 5.5.1 to 5.6.0.
- [Release notes](https://github.com/mikepenz/action-junit-report/releases)
- [Commits](https://github.com/mikepenz/action-junit-report/compare/cf701569b05ccdd861a76b8607a66d76f6fd4857...65fe03598d8d251738592a497a9e8547a5c48eaa)

---
updated-dependencies:
- dependency-name: mikepenz/action-junit-report
  dependency-version: 5.6.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependa
[…]
