+++ .github/workflows/tests-windows.yaml
@@ -118,7 +118,7 @@ jobs:
           cmake --build build --config ${{matrix.config}}
           ctest --test-dir build --build-config ${{matrix.config}} --output-on-failure --timeout 10 --output-junit tests.xml
       - name: Publish Test Report
-        uses: mikepenz/action-junit-report@cf701569b05ccdd861a76b8607a66d76f6fd4857  # v5.5.1
+        uses: mikepenz/action-junit-report@65fe03598d8d251738592a497a9e8547a5c48eaa  # v5.6.0
         if: ( success() || failure() )  # always run even if the previous step fails
         with:
           report_paths: '${{github.workspace}}/build/tests.xml'
