+++ .github/workflows/tests-ubuntu.yaml
@@ -376,7 +376,7 @@ jobs:
           else echo "with_report=false" >> "$GITHUB_OUTPUT"
           fi
       - name: Publish Test Report
-        uses: mikepenz/action-junit-report@cf701569b05ccdd861a76b8607a66d76f6fd4857  # v5.5.1
+        uses: mikepenz/action-junit-report@65fe03598d8d251738592a497a9e8547a5c48eaa  # v5.6.0
         if: ( success() || failure() ) && steps.test.outputs.with_report == 'true'  # always run even if the previous step fails
         with:
           report_paths: '${{github.workspace}}/tests.xml'
@@ -475,7 +475,7 @@ jobs:
           else echo "with_report=false" >> "$GITHUB_OUTPUT"
           fi
       - name: Publish Test Report
-        uses: mikepenz/action-junit-report@cf701569b05ccdd861a76b8607a66d76f6fd4857  # v5.5.1
+        uses: mikepenz/action-junit-report@65fe03598d8d251738592a497a9e8547a5c48eaa  # v5.6.0
         if: ( success() || failure() ) && steps.test.outputs.with_report == 'true'  # always run even if the previous step fails
         with:
           report_paths: '${{github.workspace}}/tests.xml'
