### 2025-06-05 e913f95
Release version 0.8.0 (#866)
