+++ .github/workflows/general-checks.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
       - name: Spell Check Repo
-        uses: crate-ci/typos@0f0ccba9ed1df83948f0c15026e4f5ccfce46109  # v1.32.0
+        uses: crate-ci/typos@b1ae8d918b6e85bd611117d3d9a3be4f903ee5e4  # v1.33.1
 
   trailing-whitespaces-check:
     name: Trailing whitespaces check
