### 2025-06-15 fdca66c
Add C++23 Windows and macOS build
