### 2025-06-15 2e5b674
Add C++23 msvc build
