### 2025-06-15 96d58fe
Update macOS workflow (#870)
