### 2025-06-16 c2406ff
Bump mikepenz/action-junit-report from 5.6.0 to 5.6.1
Bumps [mikepenz/action-junit-report](https://github.com/mikepenz/action-junit-report) from 5.6.0 to 5.6.1.
- [Release notes](https://github.com/mikepenz/action-junit-report/releases)
- [Commits](https://github.com/mikepenz/action-junit-report/compare/65fe03598d8d251738592a497a9e8547a5c48eaa...a83fd2b5d58d4fc702e690c1ea688d702d28d281)

---
updated-dependencies:
- dependency-name: mikepenz/action-junit-report
  dependency-version: 5.6.1
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependa
[…]
