### 2025-06-19 1ca92ea
Add an assertion to the discrete_space function and a death test
