### 2025-06-20 b5a72c8
Add comment and add a check for host_discrete_space
