+++ tests/discrete_space.cpp
@@ -25,6 +25,17 @@ TEST(DiscreteSpace, IsDiscreteSpaceInitialized)
     EXPECT_TRUE(ddc::is_discrete_space_initialized<DDimX>());
 }
 
+TEST(DiscreteSpace, HostDiscreteSpace)
+{
+#if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
+    EXPECT_DEATH(
+            ddc::host_discrete_space<DDimX>(),
+            R"rgx([Aa]ssert.*is_discrete_space_initialized<DDim>\(\))rgx");
+#else
+    GTEST_SKIP();
+#endif
+}
+
 TEST(DiscreteSpace, DiscreteSpace)
 {
 #if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
