### 2025-06-20 6584eaf
Apply suggestions from code review
Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
