### 2025-06-20 771c1bb
Add an assertion to the discrete_space function and a death test (#875)
* Add an assertion to the discrete_space function and a death test

* Add comment and add a check for host_discrete_space

* Apply suggestions from code review

Co-authored-by: Emily Bourne <louise.bourne@gmail.com>

---------

Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
