+++ namespaceddc.js
@@ -159,8 +159,8 @@ var namespaceddc =
     [ "init_discrete_space", "namespaceddc.html#a70f12a43ed0afbc092f4e974b024db3d", null ],
     [ "init_discrete_space", "namespaceddc.html#ae8ae1bc0a28343928b76730f03437c68", null ],
     [ "init_discrete_space", "namespaceddc.html#af7b4091f18fe6f867e2dea266128d8f9", null ],
-    [ "discrete_space", "namespaceddc.html#ad818604dd82f3da545fa9aebdf9168a0", null ],
     [ "is_discrete_space_initialized", "namespaceddc.html#a25c1a75897511fcd2f79f00db19fc85a", null ],
+    [ "discrete_space", "namespaceddc.html#ad818604dd82f3da545fa9aebdf9168a0", null ],
     [ "host_discrete_space", "namespaceddc.html#a7d46db599cc0436ad0b042efd27dc055", null ],
     [ "get", "namespaceddc.html#a77a97ec2669a217a5300c8c82263163a", null ],
     [ "get", "namespaceddc.html#a6141d17f89f3cc72b0472729a6aa3a54", null ],
