### 2025-06-20 726bf79
Test more recent cmake
