### 2025-06-20 2500491
Use deep_copy with an execution space
