+++ .github/workflows/tests-macos.yaml
@@ -82,61 +82,61 @@ jobs:
     steps:
       - name: Checkout built branch
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
-        with:
-          repository: pdidev/pdi
-          ref: 1.9.2
-          path: pdi
-      - name: Install PDI+user code plugin and dependencies
-        run: |
-          cmake \
-            -D BUILD_BENCHMARKING=OFF \
-            -D BUILD_DECL_HDF5_PLUGIN=OFF \
-            -D BUILD_DECL_NETCDF_PLUGIN=OFF \
-            -D BUILD_DEISA_PLUGIN=OFF \
-            -D BUILD_DOCUMENTATION=OFF \
-            -D BUILD_FORTRAN=OFF \
-            -D BUILD_MPI_PLUGIN=OFF \
-            -D BUILD_PYCALL_PLUGIN=OFF \
-            -D BUILD_SERIALIZE_PLUGIN=OFF \
-            -D BUILD_SET_VALUE_PLUGIN=OFF \
-            -D BUILD_TESTING=OFF \
-            -D BUILD_TRACE_PLUGIN=OFF \
-            -D BUILD_USER_CODE_PLUGIN=ON \
-            -D CMAKE_CXX_FLAGS="-Wno-unqualified-std-cast-call" \
-            -D CMAKE_CXX_STANDARD=17 \
-            -B build \
-            -S ./pdi
-          cmake --build build
-          cmake --install build --prefix $PDI_ROOT
-          rm -rf build
-      - name: Install fftw
-        run: brew install fftw
-      - name: Install lapack
-        run: brew install lapack
+      # - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      #   with:
+      #     repository: pdidev/pdi
+      #     ref: 1.9.2
+      #     path: pdi
+      # - name: Install PDI+user code plugin and dependencies
+      #   run: |
+      #     cmake \
+      #       -D BUILD_BENCHMARKING=OFF \
+      #       -D BUILD_DECL_HDF5_PLUGIN=OFF \
+      #       -D BUILD_DECL_NETCDF_PLUGIN=OFF \
+      #       -D BUILD_DEISA_PLUGIN=OFF \
+      #       -D BUILD_DOCUMENTATION=OFF \
+      #       -D BUILD_FORTRAN=OFF \
+      #       -D BUILD_MPI_PLUGIN=OFF \
+      #       -D BUILD_PYCALL_PLUGIN=OFF \
+      #       -D BUILD_SERIALIZE_PLUGIN=OFF \
+      #       -D BUILD_SET_VALUE_PLUGIN=OFF \
+      #       -D BUILD_TESTING=OFF \
+      #       -D BUILD_TRACE_PLUGIN=OFF \
+      #       -D BUILD_USER_CODE_PLUGIN=ON \
+      #       -D CMAKE_CXX_FLAGS="-Wno-unqualified-std-cast-call" \
+      #       -D CMAKE_CXX_STANDARD=17 \
+      #       -B build \
+      #       -S ./pdi
+      #     cmake --build build
+      #     cmake --install build --prefix $PDI_ROOT
+      #     rm -rf build
+      # - name: Install fftw
+      #   run: brew install fftw
+      # - name: Install lapack
+      #   run: brew install lapack
       - name: Install Google Test
         run: brew install googletest
       - name: Install Google benchmark
         run: brew install google-benchmark
[…]
