### 2025-06-23 daa8245
Add spline compatibility check helper
