### 2025-06-23 e47e83d
Do coverage test in NDEBUG
