### 2025-06-23 e5a1bf8
Increase discrete_space.hpp coverage (#877)
