+++ tests/discrete_space.cpp
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: MIT
 
+#include <stdexcept>
+
 #include <ddc/ddc.hpp>
 
 #include <gtest/gtest.h>
@@ -15,34 +17,29 @@ struct DDimX : ddc::UniformPointSampling<DimX>
 
 } // namespace anonymous_namespace_workaround_discrete_space_cpp
 
-TEST(DiscreteSpace, IsDiscreteSpaceInitialized)
+TEST(DiscreteSpace, Initialization)
 {
     EXPECT_FALSE(ddc::is_discrete_space_initialized<DDimX>());
-    ddc::init_discrete_space<DDimX>(DDimX::template init<DDimX>(
-            ddc::Coordinate<DimX>(0),
-            ddc::Coordinate<DimX>(1),
-            ddc::DiscreteVector<DDimX>(2)));
-    EXPECT_TRUE(ddc::is_discrete_space_initialized<DDimX>());
-}
-
-TEST(DiscreteSpace, HostDiscreteSpace)
-{
 #if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
     EXPECT_DEATH(
             ddc::host_discrete_space<DDimX>(),
             R"rgx([Aa]ssert.*is_discrete_space_initialized<DDim>\(\))rgx");
-#else
-    GTEST_SKIP();
-#endif
-}
-
-TEST(DiscreteSpace, DiscreteSpace)
-{
-#if !defined(NDEBUG) // The assertion is only checked if NDEBUG isn't defined
     EXPECT_DEATH(
             ddc::discrete_space<DDimX>(),
             R"rgx([Aa]ssert.*is_discrete_space_initialized<DDim>\(\))rgx");
-#else
-    GTEST_SKIP();
 #endif
+    ddc::init_discrete_space<DDimX>(DDimX::template init<DDimX>(
+            ddc::Coordinate<DimX>(0),
+            ddc::Coordinate<DimX>(1),
+            ddc::DiscreteVector<DDimX>(2)));
+    EXPECT_THROW(
+            ddc::init_discrete_space<DDimX>(DDimX::template init<DDimX>(
+                    ddc::Coordinate<DimX>(0),
+                    ddc::Coordinate<DimX>(1),
+                    ddc::DiscreteVector<DDimX>(2))),
+            std::runtime_error);
+    ASSERT_TRUE(ddc::is_discrete_space_initialized<DDimX>());
+    EXPECT_EQ(
+            static_cast<void const*>(&ddc::discrete_space<DDimX>()),
+            static_cast<void const*>(&ddc::host_discrete_space<DDimX>()));
 }
