+++ include/ddc/kernels/splines/spline_traits.hpp
@@ -33,7 +33,11 @@ template <
         SplineSolver Solver>
 struct is_spline_builder<SplineBuilder<ExecSpace, MemorySpace, BSplines, InterpolationDDim, BcLower, BcUpper, Solver>> : std::true_type {};
 
-template <typename T>
+/**
+ *  @brief A helper to check if T is a SplineBuilder
+ *  @tparam T The type to be checked if is a SplineBuilder
+ */
+template <class T>
 inline constexpr bool is_spline_builder_v = is_spline_builder<T>::value;
 
 template <class T>
@@ -53,7 +57,11 @@ template <
         ddc::SplineSolver Solver>
 struct is_spline_builder2D<SplineBuilder2D<ExecSpace, MemorySpace, BSpline1, BSpline2, DDimI1, DDimI2, BcLower1, BcUpper1, BcLower2, BcUpper2, Solver>> : std::true_type {};
 
-template <typename T>
+/**
+ *  @brief A helper to check if T is a SplineBuilder2D
+ *  @tparam T The type to be checked if is a SplineBuilder2D
+ */
+template <class T>
 inline constexpr bool is_spline_builder2D_v = is_spline_builder2D<T>::value;
 
 template <class T>
@@ -68,7 +76,11 @@ template <
         class UpperExtrapolationRule>
 struct is_spline_evaluator<SplineEvaluator<ExecSpace, MemorySpace, BSplines, EvaluationDDim, LowerExtrapolationRule, UpperExtrapolationRule>> : std::true_type {};
 
-template <typename T>
+/**
+ *  @brief A helper to check if T is a SplineEvaluator
+ *  @tparam T The type to be checked if is a SplineEvalutor
+ */
+template <class T>
 inline constexpr bool is_spline_evaluator_v = is_spline_evaluator<T>::value;
 
 template <class T>
@@ -87,33 +99,54 @@ template <
         class UpperExtrapolationRule2>
 struct is_spline_evaluator2D<SplineEvaluator2D<ExecSpace, MemorySpace, BSpline1, BSpline2, EvaluationDDim1, EvaluationDDim2, LowerExtrapolationRule1, UpperExtrapolationRule1, LowerExtrapolationRule2, UpperExtrapolationRule2>> : std::true_type {};
 
-template <typename T>
+/**
+ *  @brief A helper to check if T is a SplineEvaluator2D
+ *  @tparam T The type to be checked if is a SplineEvalutor2D
+ */
+template <class T>
 inline constexpr bool is_spline_evaluator2D_v = is_spline_evaluator2D<T>::value;
 
-template <class Builder, class Evaluator, class Enable = void>
-struct is_spline_compatible : std::false_type {};
-
 template <class Builder, class Evaluator>
-struct is_spline_compatible<Builder, Evaluator,
-       std::enable_if_t<is_spline_builder_v<Builder> && is_spline_evaluator_v<Evaluator>
-                        || is_spline_builder_v<Evaluator> && is_spline_evaluator_v<Builder> >> {
-       static constexpr bool value = std::is_same_v<typename Builder::exec_space, typename Evaluator::exec_space> && 
-                                     std::is_same_v<typename Builder::memory_space, typename Evaluator::memory_space> &&
-                                     std::is_same_v<typename Builder::bsplines_type, typename Evaluator::bsplines_type>;
-};
+struct is_evaluator_admissible : std::false_type {};
 
-template <class Builder, class Evaluator>
-struct is_spline_compatible<Builder, Evaluator,
-       std::enable_if_t<is_spline_builder2D_v<Builder> && is_spline_evaluator2D_v<Evaluator>
-                        || is_spline_builder2D_v<Evaluator> && is_spline_evaluator2D_v<Builder> >> {
[…]
