+++ include/ddc/kernels/splines/spline_traits.hpp
@@ -14,6 +14,7 @@
 #include <utility>
 
 #include <ddc/ddc.hpp>
+
 #include "spline_builder.hpp"
 #include "spline_builder_2d.hpp"
 #include "spline_evaluator.hpp"
@@ -21,7 +22,9 @@
 
 namespace ddc {
 template <class T>
-struct is_spline_builder : std::false_type {};
+struct is_spline_builder : std::false_type
+{
+};
 
 template <
         class ExecSpace,
@@ -31,7 +34,16 @@ template <
         ddc::BoundCond BcLower,
         ddc::BoundCond BcUpper,
         SplineSolver Solver>
-struct is_spline_builder<SplineBuilder<ExecSpace, MemorySpace, BSplines, InterpolationDDim, BcLower, BcUpper, Solver>> : std::true_type {};
+struct is_spline_builder<SplineBuilder<
+        ExecSpace,
+        MemorySpace,
+        BSplines,
+        InterpolationDDim,
+        BcLower,
+        BcUpper,
+        Solver>> : std::true_type
+{
+};
 
 /**
  *  @brief A helper to check if T is a SplineBuilder
@@ -41,7 +53,9 @@ template <class T>
 inline constexpr bool is_spline_builder_v = is_spline_builder<T>::value;
 
 template <class T>
-struct is_spline_builder2D : std::false_type {};
+struct is_spline_builder2D : std::false_type
+{
+};
 
 template <
         class ExecSpace,
@@ -55,7 +69,20 @@ template <
         ddc::BoundCond BcLower2,
         ddc::BoundCond BcUpper2,
         ddc::SplineSolver Solver>
-struct is_spline_builder2D<SplineBuilder2D<ExecSpace, MemorySpace, BSpline1, BSpline2, DDimI1, DDimI2, BcLower1, BcUpper1, BcLower2, BcUpper2, Solver>> : std::true_type {};
+struct is_spline_builder2D<SplineBuilder2D<
+        ExecSpace,
+        MemorySpace,
+        BSpline1,
+        BSpline2,
+        DDimI1,
+        DDimI2,
+        BcLower1,
+        BcUpper1,
+        BcLower2,
+        BcUpper2,
+        Solver>> : std::true_type
+{
+};
 
 /**
  *  @brief A helper to check if T is a SplineBuilder2D
@@ -65,7 +92,9 @@ template <class T>
 inline constexpr bool is_spline_builder2D_v = is_spline_builder2D<T>::value;
 
 template <class T>
-struct is_spline_evaluator : std::false_type {};
+struct is_spline_evaluator : std::false_type
+{
+};
 
 template <
         class ExecSpace,
@@ -74,7 +103,15 @@ template <
         class EvaluationDDim,
         class LowerExtrapolationRule,
         class UpperExtrapolationRule>
-struct is_spline_evaluator<SplineEvaluator<ExecSpace, MemorySpace, BSplines, EvaluationDDim, LowerExtrapolationRule, UpperExtrapolationRule>> : std::true_type {};
+struct is_spline_evaluator<SplineEvaluator<
+        ExecSpace,
+        MemorySpace,
+        BSplines,
+        EvaluationDDim,
+        LowerExtrapolationRule,
+        UpperExtrapolationRule>> : std::true_type
+{
+};
 
 /**
  *  @brief A helper to check if T is a SplineEvaluator
@@ -84,7 +121,9 @@ template <class T>
 inline constexpr bool is_spline_evaluator_v = is_spline_evaluator<T>::value;
 
 template <class T>
-struct is_spline_evaluator2D : std::false_type {};
+struct is_spline_evaluator2D : std::false_type
+{
+};
 
 template <
[…]
