+++ tests/splines/spline_traits.cpp
@@ -2,6 +2,10 @@
 //
 // SPDX-License-Identifier: MIT
 
+#include <cstddef>
+#include <type_traits>
+#include <utility>
+
 #include <ddc/ddc.hpp>
 #include <ddc/kernels/splines.hpp>
 
