### 2025-06-24 bb71594
Add spline compatibility check helper (#878)
* Add spline compatibility check helper

* rename and add docstrings

* Fix: typo

* Format

* cleanup unused headers

* add headers in tests

---------

Co-authored-by: Thomas Padioleau <thomas.padioleau@cea.fr>
Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
