+++ tests/splines/spline_traits.cpp
@@ -1,0 +1,298 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include <cstddef>
+#include <type_traits>
+#include <utility>
+
+#include <ddc/ddc.hpp>
+#include <ddc/kernels/splines.hpp>
+
+#include <gtest/gtest.h>
+
+#include <Kokkos_Core.hpp>
+
+#include "test_utils.hpp"
+
+inline namespace anonymous_namespace_workaround_spline_traits_cpp {
+
+struct DimX
+{
+    static constexpr bool PERIODIC = true;
+};
+
+struct DDimX : ddc::NonUniformPointSampling<DimX>
+{
+};
+
+struct DimY
+{
+    static constexpr bool PERIODIC = true;
+};
+
+struct DDimY : ddc::NonUniformPointSampling<DimY>
+{
+};
+
+template <typename T>
+struct BSplinesTraits;
+
+template <typename ExecSpace1, std::size_t D1, typename ExecSpace2, std::size_t D2>
+struct BSplinesTraits<std::tuple<
+        ExecSpace1,
+        std::integral_constant<std::size_t, D1>,
+        ExecSpace2,
+        std::integral_constant<std::size_t, D2>>> : public ::testing::Test
+{
+    using execution_space1 = ExecSpace1;
+    using execution_space2 = ExecSpace2;
+    using memory_space1 = typename ExecSpace1::memory_space;
+    using memory_space2 = typename ExecSpace2::memory_space;
+    static constexpr std::size_t m_spline_degree1 = D1;
+    static constexpr std::size_t m_spline_degree2 = D2;
+
+    struct BSplinesX1 : ddc::UniformBSplines<DimX, D1>
+    {
+    };
+
+    struct BSplinesX2 : ddc::UniformBSplines<DimX, D2>
+    {
+    };
+
+    struct BSplinesY : ddc::UniformBSplines<DimY, D1>
+    {
+    };
+
+    using Builder1D_1 = ddc::SplineBuilder<
+            execution_space1,
+            memory_space1,
+            BSplinesX1,
+            DDimX,
+            ddc::BoundCond::PERIODIC,
+            ddc::BoundCond::PERIODIC,
+            ddc::SplineSolver::LAPACK>;
+
+    using Evaluator1D_1 = ddc::SplineEvaluator<
+            execution_space1,
+            memory_space1,
+            BSplinesX1,
+            DDimX,
+            ddc::PeriodicExtrapolationRule<DimX>,
+            ddc::PeriodicExtrapolationRule<DimX>>;
+
+    using Builder1D_2 = ddc::SplineBuilder<
+            execution_space2,
+            memory_space2,
+            BSplinesX2,
+            DDimX,
+            ddc::BoundCond::PERIODIC,
+            ddc::BoundCond::PERIODIC,
+            ddc::SplineSolver::LAPACK>;
+
+    using Evaluator1D_2 = ddc::SplineEvaluator<
+            execution_space2,
+            memory_space2,
+            BSplinesX2,
+            DDimX,
+            ddc::PeriodicExtrapolationRule<DimX>,
+            ddc::PeriodicExtrapolationRule<DimX>>;
+
+    using Builder2D_1 = ddc::SplineBuilder2D<
+            execution_space1,
+            memory_space1,
+            BSplinesX1,
+            BSplinesY,
[…]
