### 2025-07-07 851b3dc
Bump mikepenz/action-junit-report from 5.6.1 to 5.6.2
Bumps [mikepenz/action-junit-report](https://github.com/mikepenz/action-junit-report) from 5.6.1 to 5.6.2.
- [Release notes](https://github.com/mikepenz/action-junit-report/releases)
- [Commits](https://github.com/mikepenz/action-junit-report/compare/a83fd2b5d58d4fc702e690c1ea688d702d28d281...3585e9575db828022551b4231f165eb59a0e74e3)

---
updated-dependencies:
- dependency-name: mikepenz/action-junit-report
  dependency-version: 5.6.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependa
[…]
