### 2025-07-08 646cb88
Update minimum required of kokkos-fft (#880)
