+++ README.md
@@ -47,7 +47,7 @@ To use DDC core, one needs the following dependencies:
 To use DDC components, one needs the following dependencies:
 
 * (optional, fft interface) DDC::fft
-  * Kokkos-fft 0.2.1...<1
+  * Kokkos-fft 0.3.0...<1
 * (optional, IO interface) DDC::pdi
   * PDI 1.6...<2
 * (optional, spline interpolation) DDC::splines
