+++ cmake/DDCConfig.cmake.in
@@ -42,7 +42,7 @@ if(NOT TARGET DDC::core)
 endif()
 
 if(@DDC_BUILD_KERNELS_FFT@ AND ("fft" IN_LIST DDC_FIND_COMPONENTS))
-   ddc_find_dependency(KokkosFFT 0.2.1...<1)
+   ddc_find_dependency(KokkosFFT 0.3.0...<1)
 
    include(${CMAKE_CURRENT_LIST_DIR}/DDCFftTargets.cmake)
    if(NOT TARGET DDC::fft)
