### 2025-07-08 44ae8ea
Try -undefined dynamic_lookup flags on macOS
