+++ .github/workflows/tests-macos.yaml
@@ -121,12 +121,13 @@ jobs:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
         with:
           repository: ginkgo-project/ginkgo
-          ref: v1.9.0
+          ref: v1.8.0
           path: ginkgo
       - name: Install Ginkgo
         run: |
           cmake \
             -D CMAKE_CXX_STANDARD=${{matrix.cxx_version}} \
+            -D CMAKE_CXX_FLAGS="-undefined dynamic_lookup" \
             -D GINKGO_BUILD_BENCHMARKS=OFF \
             -D GINKGO_BUILD_EXAMPLES=OFF \
             -D GINKGO_BUILD_MPI=OFF \
