### 2025-07-08 60be61d
Pass only the flag to shared library linker
