### 2025-07-08 873eb73
Explicit use of macos-15 (#887)
* Explicit use of macos-15

* Use -undefined dynamic_lookup flags on macOS
