### 2025-07-08 5796f18
Remove last front call
