### 2025-07-09 1512911
Add macros test
