+++ install_test/check_macros/check_macros.py
@@ -1,0 +1,200 @@
+#!/usr/bin/env python3
+
+# Copyright (C) The DDC development team, see COPYRIGHT.md file
+#
+# SPDX-License-Identifier: MIT
+
+"""
+Compare the difference in preprocessor macros between two source files
+(one with a library and one without), and check if that diff matches a reference.
+
+Uses compile_commands.json to simulate compilation and extract preprocessor macros.
+The reference file must be a JSON file containing a list of expected #define macros.
+"""
+
+import json
+import subprocess
+import shlex
+from pathlib import Path
+import argparse
+import sys
+
+
+def extract_command(entry):
+    """
+    Modify compile command to only run preprocessor: remove -c/-o, add -dM -E.
+
+    Args:
+        entry (dict): An entry from compile_commands.json
+
+    Returns:
+        tuple[list[str], Path]: The modified command and working directory
+    """
+    parts = shlex.split(entry["command"])
+    new_parts = []
+    skip_next = False
+    for part in parts:
+        if skip_next:
+            skip_next = False
+            continue
+        if part == "-c":
+            continue
+        if part == "-o":
+            skip_next = True
+            continue
+        new_parts.append(part)
+    new_parts += ["-dM", "-E", entry["file"]]
+    return new_parts, Path(entry["directory"])
+
+
+def run_preprocessor(command, cwd):
+    """
+    Run the preprocessor and return unique macro lines.
+
+    Args:
+        command (list[str]): Compiler command
+        cwd (Path): Working directory
+
+    Returns:
+        set[str]: Set of macro lines
+    """
+    result = subprocess.run(
+        command, cwd=cwd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, check=True
+    )
+    if result.returncode != 0:
+        raise RuntimeError(f"Command failed:\n{' '.join(command)}\n{result.stderr}")
+    return set(result.stdout.splitlines())
+
+
+def find_entry(commands, source_path):
+    """
+    Find compile command for the given source file.
+
+    Args:
+        commands (list[dict]): compile_commands.json content
+        source_path (Path): Path to source file
+
+    Returns:
+        dict: Matching compile command entry
+    """
+    for entry in commands:
+        if Path(entry["file"]) == source_path:
+            return entry
+    raise ValueError(f"Source file {source_path} not found in compile_commands.json")
+
+
+def prepare_commands(args):
+    """
+    Load compilation database and extract preprocessor commands for both source files.
+
+    Args:
+        args (argparse.Namespace): Parsed command-line arguments
+
+    Returns:
+        tuple[set[str], set[str]]: Macros with and without the library
+    """
+    commands = json.loads(args.db.read_text())
+
+    entry_with = find_entry(commands, args.with_src.resolve())
+    entry_without = find_entry(commands, args.without_src.resolve())
+
+    cmd_with, dir_with = extract_command(entry_with)
+    cmd_without, dir_without = extract_command(entry_without)
+
+    print(f"Running preprocessor for: {args.with_src.name}")
+    macros_with = run_preprocessor(cmd_with, dir_with)
+
+    print(f"Running preprocessor for: {args.without_src.name}")
+    macros_without = run_preprocessor(cmd_without, dir_without)
+
+    return macros_with, macros_without
+
+
+def extract_macro_diff(macros_with, macros_without):
+    """
+    Compute the difference in #define macros between with/without source.
+
+    Args:
+        macros_with (set[str]): Macros when library is included
+        macros_without (set[str]): Macros without library
+
+    Returns:
+        list[str]: Sorted list of macro differences
+    """
+    define_with = {
+        line.removeprefix("#define").strip() for line in macros_with if line.startswith("#define")
+    }
[…]
