### 2025-07-09 0f374ad
Add macros test
