+++ .github/workflows/tests-ubuntu.yaml
@@ -389,6 +389,21 @@ jobs:
               -B build \
               -S /src/install_test
             cmake --build build
+            rm -rf build
+
+            if [ 'xcpu' = 'x${{matrix.backend.name}}' ]
+            then
+              cmake \
+                -D CMAKE_EXPORT_COMPILE_COMMANDS=ON \
+                -B build \
+                -S /src/install_test/check_macros
+              python3 \
+                /src/install_test/check_macros/check_macros.py \
+                --db ./build/compile_commands.json \
+                --with-src /src/install_test/check_macros/with_my_lib.cpp \
+                --without-src /src/install_test/check_macros/without_my_lib.cpp \
+                --reference /src/install_test/check_macros/whitelist_macros.json
+            fi
           EOF
 
           docker run \
