### 2025-07-09 f17c6a3
Add macros test
