### 2025-07-09 d08aad0
Add macros test
