### 2025-07-09 f4eb414
Add macros test
