### 2025-07-09 8be42c1
Add macros test
