### 2025-07-12 0c112d9
Add Python in dockerfiles (#890)
