### 2025-07-12 8513155
Remove macro definition
