### 2025-07-12 fba358c
Remove macro definition
