### 2025-07-12 c25a27a
Remove macro definition
