### 2025-07-12 9940dd3
Remove macro definition
