### 2025-07-12 75d7508
Remove macro definition
