### 2025-07-13 65f5078
Fix test for CUDA and HIP
