### 2025-07-13 9a88b0f
Fix test for CUDA and HIP
