### 2025-07-13 d810335
Tolerate some macros
