### 2025-07-13 c67d5ba
Fix CUDA compilation
