### 2025-07-13 ec83355
Fix CUDA compilation
