### 2025-07-13 aaac4ca
Fix CUDA compilation
