### 2025-07-13 057c406
Fix CUDA compilation
