### 2025-07-13 068ac80
Fix CUDA compilation
