+++ install_test/check_macros/without_my_lib.cpp
@@ -3,8 +3,9 @@
 // SPDX-License-Identifier: MIT
 
 #include <any>
-#include <iostream>
+#include <filesystem>
 #include <list>
+#include <ranges>
 
 #include <KokkosFFT.hpp>
 #include <Kokkos_Core.hpp>
