### 2025-07-18 799f757
Add missing strides() method. Allow conversion from DiscreteDomain to StridedDiscreteDomain
