### 2025-07-18 02fb8e2
Fix slice of ChunkSpan on StridedDiscreteDomain. Fixes #891
