### 2025-07-22 efde399
Fix slicing operator
