### 2025-07-22 7319f84
Avoid mixing ADL and RebindDomain
