### 2025-07-22 9fb1d13
Fix slicing operator (#893)
* Fix slicing operator

* Avoid mixing ADL and RebindDomain
