+++ examples/characteristics_advection.cpp
@@ -254,8 +254,9 @@ int main(int argc, char** argv)
                     feet_coords(e) = ddc::coordinate(ddc::DiscreteElement<DDimX>(e))
                                      - ddc::Coordinate<X>(vx * ddc::step<DDimT>());
                 });
+        ddc::ChunkSpan<double const, ddc::StridedDiscreteDomain<ddc::Deriv<X>, DDimX, DDimY>> derivs;
         // Interpolate the values at feet on the grid
-        spline_builder(coef, last_density.span_cview());
+        spline_builder(coef, last_density.span_cview(), derivs);
         spline_evaluator(next_density, feet_coords.span_cview(), coef.span_cview());
         //! [numerical scheme]
 
