+++ .github/workflows/general-checks.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
       - name: Spell Check Repo
-        uses: crate-ci/typos@392b78fe18a52790c53f42456e46124f77346842  # v1.34.0
+        uses: crate-ci/typos@52bd719c2c91f9d676e2aa359fc8e0db8925e6d8  # v1.35.3
 
   trailing-whitespaces-check:
     name: Trailing whitespaces check
