+++ .github/workflows/pages.yaml
@@ -88,7 +88,7 @@ jobs:
           submodules: true
       - name: Collect image artifact
         if: needs.id_repo.outputs.in_base_repo == 'false'
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093  # v4.3.0
+        uses: actions/download-artifact@634f93cb2916e3fdff6788551b99b062d0335ce0  # v5.0.0
         with: {name: 'doxygen-artifact'}
       - name: Load image artifact into docker
         if: needs.id_repo.outputs.in_base_repo == 'false'
@@ -149,7 +149,7 @@ jobs:
           submodules: true
       - name: Collect image artifact
         if: needs.id_repo.outputs.in_base_repo == 'false'
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093  # v4.3.0
+        uses: actions/download-artifact@634f93cb2916e3fdff6788551b99b062d0335ce0  # v5.0.0
         with: {name: 'doxygen-artifact'}
       - name: Load image artifact into docker
         if: needs.id_repo.outputs.in_base_repo == 'false'
