+++ .github/workflows/tests-ubuntu.yaml
@@ -186,7 +186,7 @@ jobs:
             ghcr.io/cexa-project/ddc/${{matrix.image}}_${{matrix.backend.name}}:${GITHUB_SHA:0:7} \
             bash run.sh
       - name: Upload coverage reports to Codecov with GitHub Action
-        uses: codecov/codecov-action@18283e04ce6e62d37312384ff67231eb8fd56d24  # v5.4.3
+        uses: codecov/codecov-action@fdcc8476540edceab3de004e990f80d881c6cc00  # v5.5.0
         with:
           files: coverage.lcov
           disable_search: true
