### 2025-08-25 ac321c0
Update .github/workflows/scorecard.yaml
