+++ .github/workflows/python-checks.yaml
@@ -29,7 +29,7 @@ jobs:
     name: Python format using black
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
       - name: Install dependencies
         run: |
           pipx install black
@@ -41,7 +41,7 @@ jobs:
     name: Python lint using pylint
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
       - name: Install dependencies
         run: |
           pipx install pylint
