+++ .github/workflows/cmake-checks.yaml
@@ -33,7 +33,7 @@ jobs:
     name: CMake format check using gersemi
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
       - name: Install gersemi
         run: pipx install gersemi~=0.19
       - name: Run gersemi
