+++ .github/workflows/tests-windows.yaml
@@ -67,8 +67,8 @@ jobs:
       checks: write  # Required by mikepenz/action-junit-report
     steps:
       - name: Checkout built branch
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           repository: google/googletest
           ref: v1.16.0
@@ -84,7 +84,7 @@ jobs:
           cmake --build build --config ${{matrix.config}}
           cmake --install build --config ${{matrix.config}} --prefix $GTest_ROOT
           rm -rf build
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           repository: kokkos/kokkos
           ref: 4.6.00
