+++ .github/workflows/general-checks.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
       - name: Spell Check Repo
-        uses: crate-ci/typos@52bd719c2c91f9d676e2aa359fc8e0db8925e6d8  # v1.35.3
+        uses: crate-ci/typos@a4c3e43aea0a9e9b9e6578d2731ebd9a27e8f6cd  # v1.35.5
 
   trailing-whitespaces-check:
     name: Trailing whitespaces check
