### 2025-08-26 fbbce3f
Bump github/codeql-action from 3.29.8 to 3.29.11 (#904)
* Bump github/codeql-action from 3.29.8 to 3.29.11

Bumps [github/codeql-action](https://github.com/github/codeql-action) from 3.29.8 to 3.29.11.
- [Release notes](https://github.com/github/codeql-action/releases)
- [Changelog](https://github.com/github/codeql-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/github/codeql-action/compare/76621b61decf072c1cee8dd1ce2d2a82d33c17ed...3c3833e0f8c1c83d449a7478aa59c036a9165498)

---
updated-dependencies:
- dependency-name: github/codeql-action
  dependency-version: 3.29.11
  dependen
[…]
