+++ .github/workflows/website-checks.yaml
@@ -19,7 +19,7 @@ jobs:
     name: Links check using linkchecker
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+      - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           ref: gh-pages
       - name: Install linkchecker
