### 2025-08-28 #8256 @danielskatz
@EmilyBourne - I apologize, but this is far behind the normal JOSS schedule, and I can't predict when it will make it through.
