### 2025-08-28 #8256 @EmilyBourne
> [@EmilyBourne](https://github.com/EmilyBourne) - I apologize, but this is far behind the normal JOSS schedule, and I can't predict when it will make it through.

Thank you for your speedy reply. In that case we will not wait for this paper.
