### 2025-09-01 76191c8
Remove C++17 builds in the early integration (#908)
