### 2025-09-01 16c9555
Update Spack installation documentation
