+++ search/pages_4.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['getting_20started_0',['Getting started',['../md_getting__started.html',1,'']]]
+  ['getting_20started_0',['Getting started',['../md_getting__started.html',1,'']]],
+  ['guide_1',['Installation guide',['../installation.html',1,'md_getting__started']]]
 ];
