+++ search/pages_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['installation_0',['Installation',['../installation.html',1,'md_getting__started']]]
+  ['installation_20guide_0',['Installation guide',['../installation.html',1,'md_getting__started']]]
 ];

[…]
