### 2025-09-02 #8256 @danielskatz
You can add the second reviewer and start the review, but remember/note that the second reviewer won't start until then. This lets the first reviewer start, which can lead to improvements that might make the second review faster.
