+++ .github/workflows/tests-macos.yaml
@@ -142,7 +142,7 @@ jobs:
       - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           repository: kokkos/kokkos
-          ref: 4.6.00
+          ref: 4.7.00
           path: kokkos
       - name: Install Kokkos
         run: |
@@ -175,7 +175,7 @@ jobs:
       - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           repository: kokkos/kokkos-kernels
-          ref: 4.6.00
+          ref: 4.7.00
           path: kokkos-kernels
       - name: Install Kokkos Kernels
         run: |
