### 2025-09-03 4a9cb43
Add C++23 Windows and macOS builds (#869)
* Add C++23 Windows and macOS build

* Update Kokkos
