+++ .github/workflows/general-checks.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
       - name: Spell Check Repo
-        uses: crate-ci/typos@65f69f021b736bdbe548ce72200500752d42b40e  # v1.35.7
+        uses: crate-ci/typos@85f62a8a84f939ae994ab3763f01a0296d61a7ee  # v1.36.2
 
   trailing-whitespaces-check:
     name: Trailing whitespaces check
