+++ CMakeLists.txt
@@ -126,17 +126,54 @@ endif()
 add_library(ddc_core INTERFACE)
 add_library(DDC::core ALIAS ddc_core)
 configure_file(cmake/config.hpp.in generated/ddc/config.hpp NO_SOURCE_PERMISSIONS @ONLY)
-install(FILES ${CMAKE_CURRENT_BINARY_DIR}/generated/ddc/config.hpp DESTINATION include/ddc)
-install(TARGETS ddc_core EXPORT DDCTargets)
-target_compile_features(ddc_core INTERFACE cxx_std_17)
-target_include_directories(
+target_sources(
     ddc_core
     INTERFACE
-        "$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/generated>"
-        "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
-        "$<INSTALL_INTERFACE:${DDC_INSTALL_INCLUDEDIR}>"
+        FILE_SET HEADERS
+            BASE_DIRS include ${CMAKE_CURRENT_BINARY_DIR}/generated
+            FILES
+                ${CMAKE_CURRENT_BINARY_DIR}/generated/ddc/config.hpp
+                include/ddc/detail/dual_discretization.hpp
+                include/ddc/detail/kokkos.hpp
+                include/ddc/detail/macros.hpp
+                include/ddc/detail/tagged_vector.hpp
+                include/ddc/detail/type_seq.hpp
+                include/ddc/detail/type_traits.hpp
+                include/ddc/aligned_allocator.hpp
+                include/ddc/chunk.hpp
+                include/ddc/chunk_common.hpp
+                include/ddc/chunk_span.hpp
+                include/ddc/chunk_traits.hpp
+                include/ddc/coordinate.hpp
+                include/ddc/create_mirror.hpp
+                include/ddc/ddc.hpp
+                include/ddc/ddc_to_kokkos_execution_policy.hpp
+                include/ddc/discrete_domain.hpp
+                include/ddc/discrete_element.hpp
+                include/ddc/discrete_space.hpp
+                include/ddc/discrete_vector.hpp
+                include/ddc/for_each.hpp
+                include/ddc/kokkos_allocator.hpp
+                include/ddc/non_uniform_point_sampling.hpp
+                include/ddc/parallel_deepcopy.hpp
+                include/ddc/parallel_fill.hpp
+                include/ddc/parallel_for_each.hpp
+                include/ddc/parallel_transform_reduce.hpp
+                include/ddc/periodic_sampling.hpp
+                include/ddc/print.hpp
+                include/ddc/real_type.hpp
+                include/ddc/reducer.hpp
+                include/ddc/scope_guard.hpp
+                include/ddc/sparse_discrete_domain.hpp
+                include/ddc/strided_discrete_domain.hpp
+                include/ddc/transform_reduce.hpp
+                include/ddc/trivial_space.hpp
+                include/ddc/uniform_point_sampling.hpp
 )
+target_compile_features(ddc_core INTERFACE cxx_std_17)
 target_link_libraries(ddc_core INTERFACE Kokkos::kokkos)
+
+install(TARGETS ddc_core EXPORT DDCTargets FILE_SET HEADERS)
 
 # Link library to DDC
 
@@ -167,9 +204,16 @@ if("${DDC_BUILD_KERNELS_FFT}")
 
     add_library(ddc_fft INTERFACE)
     add_library(DDC::fft ALIAS ddc_fft)
-    install(TARGETS ddc_fft EXPORT DDCFftTargets)
[…]
