### 2025-09-15 85b3b35
Install PDI from Ubuntu repositories
