### 2025-09-15 5816a81
Install PDI from Ubuntu repositories (#918)
