### 2025-09-15 f6d8715
Fix warning about windows-latest (#916)
