### 2025-09-16 9301fc5
Run benchmarks in CI (#920)
