+++ .github/workflows/tests-ubuntu.yaml
@@ -354,7 +354,10 @@ jobs:
               ./build/examples/heat_equation_spectral
               ./build/examples/heat_equation
               ./build/examples/non_uniform_heat_equation
+              ./build/examples/pdi_event
               ./build/examples/uniform_heat_equation
+              ./build/benchmarks/ddc_benchmark_deepcopy --benchmark_dry_run
+              ./build/benchmarks/ddc_benchmark_splines --benchmark_dry_run
             fi
             cmake --install build --prefix $DDC_ROOT
             rm -rf build
@@ -461,7 +464,10 @@ jobs:
             ./build/examples/heat_equation_spectral
             ./build/examples/heat_equation
             ./build/examples/non_uniform_heat_equation
+            ./build/examples/pdi_event
             ./build/examples/uniform_heat_equation
+            ./build/benchmarks/ddc_benchmark_deepcopy --benchmark_dry_run
+            ./build/benchmarks/ddc_benchmark_splines --benchmark_dry_run
           EOF
 
           docker run \
