### 2025-09-16 b351b70
Update cmake minimum required to 3.25 and docker images
