+++ examples/CMakeLists.txt
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: MIT
 
-cmake_minimum_required(VERSION 3.22)
+cmake_minimum_required(VERSION 3.25)
 
 add_executable(game_of_life game_of_life.cpp)
 target_link_libraries(game_of_life PUBLIC DDC::core)
