### 2025-09-17 97e8db9
Update cmake minimum required to 3.25 and docker images (#921)
* Update cmake minimum required to 3.25 and docker images

* Update documentation
