### 2025-09-17 #8256 @pachadotdev
> [@pachadotdev](https://github.com/pachadotdev) Thank you very much, could we get a rough estimation when the reviews will start ?

I understand rev. 1 can start. Can you provide an estimate, @lwshanbd ?
