### 2025-09-17 #8256 @danielskatz
@pachadotdev - if you have 2 reviewers, you should start the review so that the first can start their part, as I mentioned in https://github.com/openjournals/joss-reviews/issues/8256#issuecomment-3247020895
