### 2025-09-19 97d7fd6
Use CMake file sets to install headers (#915)
* Use CMake file sets to install headers

* Update the installation test

* Fix missing headers
