### 2025-09-19 028c5ad
Improve the install test and add SYSTEM to vendored libraries
