### 2025-09-20 6994ca6
Add missing SYSTEM for Kokkos Kernels
