+++ cmake/DDCVendorConfiguration.cmake
@@ -54,5 +54,5 @@ macro(ddc_configure_kokkos_kernels)
     set(KokkosKernels_ENABLE_COMPONENT_LAPACK OFF)
     set(KokkosKernels_ENABLE_TPL_BLAS OFF)
     set(KokkosKernels_ENABLE_TPL_LAPACK OFF)
-    add_subdirectory(vendor/kokkos-kernels)
+    add_subdirectory(vendor/kokkos-kernels SYSTEM)
 endmacro()
