### 2025-09-26 4b408a3
Add a 3D spline builder for the periodic case
