+++ include/ddc/kernels/splines/spline_evaluator_3d.hpp
@@ -1,0 +1,2425 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include <array>
+#include <cstddef>
+#include <type_traits>
+
+#include <ddc/ddc.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include "integrals.hpp"
+#include "periodic_extrapolation_rule.hpp"
+
+namespace ddc {
+
+/**
+ * @brief A class to evaluate, differentiate or integrate a 3D spline function.
+ *
+ * A class which contains an operator () which can be used to evaluate, differentiate or integrate a 3D spline function.
+ *
+ * @tparam ExecSpace The Kokkos execution space on which the spline evaluation is performed.
+ * @tparam MemorySpace The Kokkos memory space on which the data (spline coefficients and evaluation) is stored.
+ * @tparam BSplines1 The discrete dimension representing the B-splines along the first dimension of interest.
+ * @tparam BSplines2 The discrete dimension representing the B-splines along the second dimension of interest.
+ * @tparam BSplines3 The discrete dimension representing the B-splines along the third dimension of interest.
+ * @tparam EvaluationDDim1 The first discrete dimension on which evaluation points are defined.
+ * @tparam EvaluationDDim2 The second discrete dimension on which evaluation points are defined.
+ * @tparam EvaluationDDim3 The third discrete dimension on which evaluation points are defined.
+ * @tparam LowerExtrapolationRule1 The lower extrapolation rule type along first dimension of interest.
+ * @tparam UpperExtrapolationRule1 The upper extrapolation rule type along first dimension of interest.
+ * @tparam LowerExtrapolationRule2 The lower extrapolation rule type along second dimension of interest.
+ * @tparam UpperExtrapolationRule2 The upper extrapolation rule type along second dimension of interest.
+ * @tparam LowerExtrapolationRule3 The lower extrapolation rule type along third dimension of interest.
+ * @tparam UpperExtrapolationRule3 The upper extrapolation rule type along third dimension of interest.
+ */
+template <
+        class ExecSpace,
+        class MemorySpace,
+        class BSplines1,
+        class BSplines2,
+        class BSplines3,
+        class EvaluationDDim1,
+        class EvaluationDDim2,
+        class EvaluationDDim3,
+        class LowerExtrapolationRule1,
+        class UpperExtrapolationRule1,
+        class LowerExtrapolationRule2,
+        class UpperExtrapolationRule2,
+        class LowerExtrapolationRule3,
+        class UpperExtrapolationRule3>
+class SplineEvaluator3D
+{
+private:
+    /**
+     * @brief Tag to indicate that the value of the spline should be evaluated.
+     */
+    struct eval_type
+    {
+    };
+
+    /**
+     * @brief Tag to indicate that derivative of the spline should be evaluated.
+     */
+    struct eval_deriv_type
+    {
+    };
+
+public:
+    /// @brief The type of the first evaluation continuous dimension used by this class.
+    using continuous_dimension_type1 = typename BSplines1::continuous_dimension_type;
+
+    /// @brief The type of the second evaluation continuous dimension used by this class.
+    using continuous_dimension_type2 = typename BSplines2::continuous_dimension_type;
+
+    /// @brief The type of the third evaluation continuous dimension used by this class.
+    using continuous_dimension_type3 = typename BSplines3::continuous_dimension_type;
+
+    /// @brief The type of the Kokkos execution space used by this class.
+    using exec_space = ExecSpace;
+
+    /// @brief The type of the Kokkos memory space used by this class.
+    using memory_space = MemorySpace;
+
+    /// @brief The type of the first discrete dimension of interest used by this class.
[…]
