+++ .github/workflows/general-checks.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
       - name: Spell Check Repo
-        uses: crate-ci/typos@85f62a8a84f939ae994ab3763f01a0296d61a7ee  # v1.36.2
+        uses: crate-ci/typos@0c17dabcee8b8f1957fa917d17393a23e02e1583  # v1.36.3
 
   trailing-whitespaces-check:
     name: Trailing whitespaces check
