+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -61,20 +61,12 @@ public:
     using builder_type3 = ddc::
             SplineBuilder<ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower3, BcUpper3, Solver>;
 
-    // FIXME: change this when implementing the derivatives
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate the second-dimension-derivatives along first dimension.
-    using builder_deriv_type1 = ddc::SplineBuilder2D<
-            ExecSpace,
-            MemorySpace,
-            BSpline1,
-            BSpline2,
-            DDimI1,
-            DDimI2,
-            BcLower1,
-            BcUpper1,
-            BcLower2,
-            BcLower2,
-            Solver>;
+    // TODO: add documentation comment
+    using builder_deriv_type1 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
+
+    using builder_deriv_type2 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
 
     /// @brief The type of the first interpolation continuous dimension.
     using continuous_dimension_type1 = typename builder_type1::continuous_dimension_type;
@@ -191,44 +183,11 @@ public:
      * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y and Z,
      * this is DiscreteDomain<Deriv<X>, Y, Z, T>.
      */
-    // template <
-    //         class BatchedInterpolationDDom,
-    //         class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    // using batched_derivs_domain_type1 =
-    //         typename builder_type1::template batched_derivs_domain_type<BatchedInterpolationDDom>;
-
     template <
             class BatchedInterpolationDDom,
             class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    using batched_derivs_domain_type_1_2 = ddc::replace_dim_of_t<
-            ddc::replace_dim_of_t<
-                    BatchedInterpolationDDom,
-                    interpolation_discrete_dimension_type1,
-                    deriv_type1>,
-            interpolation_discrete_dimension_type2,
-            deriv_type2>;
-
-    template <
-            class BatchedInterpolationDDom,
-            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    using batched_derivs_domain_type_2_3 = ddc::replace_dim_of_t<
-            ddc::replace_dim_of_t<
-                    BatchedInterpolationDDom,
-                    interpolation_discrete_dimension_type2,
-                    deriv_type2>,
-            interpolation_discrete_dimension_type3,
-            deriv_type3>;
-
-    template <
-            class BatchedInterpolationDDom,
-            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    using batched_derivs_domain_type_1_3 = ddc::replace_dim_of_t<
-            ddc::replace_dim_of_t<
-                    BatchedInterpolationDDom,
-                    interpolation_discrete_dimension_type3,
-                    deriv_type3>,
-            interpolation_discrete_dimension_type3,
-            deriv_type3>;
+    using batched_derivs_domain_type1 =
+            typename builder_type1::template batched_derivs_domain_type<BatchedInterpolationDDom>;
 
     /**
      * @brief The type of the whole Derivs domain (cartesian product of the 1D Deriv domain
[…]
