### 2025-09-30 6875f7a
Reword the FIXMEs in the 3D test
