+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -12,7 +12,6 @@
 #include <ddc/ddc.hpp>
 
 #include "spline_builder.hpp"
-#include "spline_builder_2d.hpp"
 
 namespace ddc {
 
