### 2025-09-30 6291ff3
Add missing doc comments and add unused annotations to the derivs parameters
