### 2025-10-02 11b2f8e
Modify 3D error bound computation
