+++ tests/splines/spline_error_bounds.hpp
@@ -173,9 +173,9 @@ public:
             int degree2,
             int degree3) const
     {
-        double const norm1 = m_evaluator.max_norm(degree1 + 1, 1, 1);
-        double const norm2 = m_evaluator.max_norm(1, degree2 + 1, 1);
-        double const norm3 = m_evaluator.max_norm(1, 1, degree3 + 1);
+        double const norm1 = m_evaluator.max_norm(degree1 + 1, 1, 0);
+        double const norm2 = m_evaluator.max_norm(1, degree2 + 1, 0);
+        double const norm3 = m_evaluator.max_norm(0, 0, degree3 + 1);
         return tihomirov_error_bound(cell_width1, degree1 - 1, norm1)
                + tihomirov_error_bound(cell_width2, degree2 - 1, norm2)
                + tihomirov_error_bound(cell_width3, degree3, norm3);
@@ -189,12 +189,12 @@ public:
             int degree2,
             int degree3) const
     {
-        double const norm1 = m_evaluator.max_norm(degree1 + 1, 1, 1);
-        double const norm2 = m_evaluator.max_norm(1, degree2 + 1, 1);
-        double const norm3 = m_evaluator.max_norm(1, 1, degree3 + 1);
+        double const norm1 = m_evaluator.max_norm(degree1 + 1, 0, 0);
+        double const norm2 = m_evaluator.max_norm(0, degree2 + 1, 1);
+        double const norm3 = m_evaluator.max_norm(0, 1, degree3 + 1);
         return tihomirov_error_bound(cell_width1, degree1, norm1)
                + tihomirov_error_bound(cell_width2, degree2 - 1, norm2)
-               + tihomirov_error_bound(cell_width3, degree3 - 3, norm3);
+               + tihomirov_error_bound(cell_width3, degree3 - 1, norm3);
     }
 
     double error_bound_on_deriv_13(
@@ -205,9 +205,9 @@ public:
             int degree2,
             int degree3) const
     {
-        double const norm1 = m_evaluator.max_norm(degree1 + 1, 1, 1);
-        double const norm2 = m_evaluator.max_norm(1, degree2 + 1, 1);
-        double const norm3 = m_evaluator.max_norm(1, 1, degree3 + 1);
+        double const norm1 = m_evaluator.max_norm(degree1 + 1, 0, 1);
+        double const norm2 = m_evaluator.max_norm(0, degree2 + 1, 0);
+        double const norm3 = m_evaluator.max_norm(1, 0, degree3 + 1);
         return tihomirov_error_bound(cell_width1, degree1 - 1, norm1)
                + tihomirov_error_bound(cell_width2, degree2, norm2)
                + tihomirov_error_bound(cell_width3, degree3 - 1, norm3);
