### 2025-10-02 1edaf9e
Use annotate_only in action-junit-report
