+++ .github/workflows/tests-ubuntu.yaml
@@ -236,8 +236,6 @@ jobs:
             cxx_version: '23'
     runs-on: ubuntu-latest
     needs: [docker-build, id_repo]
-    permissions:
-      checks: write  # Required by mikepenz/action-junit-report
     steps:
       - name: Free Disk Space (Ubuntu)
         uses: jlumbroso/free-disk-space@54081f138730dfa15788a46383842cd2f914a1be  # v1.3.1
@@ -381,6 +379,7 @@ jobs:
         uses: mikepenz/action-junit-report@3585e9575db828022551b4231f165eb59a0e74e3  # v5.6.2
         if: ( success() || failure() ) && steps.test.outputs.with_report == 'true'  # always run even if the previous step fails
         with:
+          annotate_only: true  # forked repo cannot write to checks so just do annotations
           report_paths: '${{github.workspace}}/tests.xml'
 
   test-with-sanitizer:
@@ -398,8 +397,6 @@ jobs:
         sanitizer: ['address', 'undefined']
     runs-on: ubuntu-latest
     needs: [docker-build, id_repo]
-    permissions:
-      checks: write  # Required by mikepenz/action-junit-report
     steps:
       - name: Free Disk Space (Ubuntu)
         uses: jlumbroso/free-disk-space@54081f138730dfa15788a46383842cd2f914a1be  # v1.3.1
@@ -483,6 +480,7 @@ jobs:
         uses: mikepenz/action-junit-report@3585e9575db828022551b4231f165eb59a0e74e3  # v5.6.2
         if: ( success() || failure() ) && steps.test.outputs.with_report == 'true'  # always run even if the previous step fails
         with:
+          annotate_only: true  # forked repo cannot write to checks so just do annotations
           report_paths: '${{github.workspace}}/tests.xml'
 
   clang-tidy-test:
