### 2025-10-02 e0e94ab
Add hermite boundary conditions
