+++ tests/splines/batched_3d_spline_builder.cpp
@@ -209,21 +209,34 @@ void Batched3dSplineTest()
             interpolation_domain3);
 
 #if defined(BC_HERMITE)
-    // FIXME: This is the 2D version, modify when adding the Hermite boundary conditions
-
     // Create the derivs domain
     ddc::DiscreteDomain<ddc::Deriv<I1>> const
             derivs_domain1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I2>> const
             derivs_domain2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
-    ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const
-            derivs_domain(derivs_domain1, derivs_domain2);
+    ddc::DiscreteDomain<ddc::Deriv<I3>> const
+            derivs_domain3(DElem<ddc::Deriv<I3>>(1), DVect<ddc::Deriv<I3>>(s_degree / 2));
+    ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>, DDimI3> const
+            derivs_domain12(derivs_domain1, derivs_domain2, interpolation_domain3);
+    ddc::DiscreteDomain<DDimI1, ddc::Deriv<I2>, ddc::Deriv<I3>> const
+            derivs_domain23(interpolation_domain1, derivs_domain2, derivs_domain3);
+    ddc::DiscreteDomain<ddc::Deriv<I1>, DDimI2, ddc::Deriv<I3>> const
+            derivs_domain13(derivs_domain1, interpolation_domain2, derivs_domain3);
+    ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>, ddc::Deriv<I3>> const
+            derivs_domain(derivs_domain1, derivs_domain2, derivs_domain3);
 
-    auto const dom_derivs_1d
-            = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_vals, derivs_domain1);
+    auto const dom_derivs1 = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_vals, derivs_domain1);
     auto const dom_derivs2 = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_vals, derivs_domain2);
-    auto const dom_derivs
-            = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs_1d, derivs_domain2);
+    auto const dom_derivs3 = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_vals, derivs_domain3);
+    auto const dom_derivs12
+            = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs1, derivs_domain2);
+    auto const dom_derivs23
+            = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs2, derivs_domain3);
+    auto const dom_derivs13
+            = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs1, derivs_domain3);
+    auto const dom_derivs = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(
+            ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs1, derivs_domain2),
+            derivs_domain3);
 #endif
 
     // Create a SplineBuilder over BSplines<I> and batched along other dimensions using some boundary conditions
@@ -267,63 +280,67 @@ void Batched3dSplineTest()
[…]
