+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -12,6 +12,9 @@
 #include <ddc/ddc.hpp>
 
 #include "spline_builder.hpp"
+#include "spline_builder_2d.hpp"
+
+#define USE_RECURSIVE_BUILDER
 
 namespace ddc {
 
@@ -67,6 +70,21 @@ public:
     /// @brief The type of the SplineBuilder used by this class to spline-approximate the third-dimension-derivatives along second dimension.
     using builder_deriv_type2 = ddc::
             SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
+
+#ifdef USE_RECURSIVE_BUILDER
+    using builder_type2d = ddc::SplineBuilder2D<
+            ExecSpace,
+            MemorySpace,
+            BSpline2,
+            BSpline3,
+            DDimI2,
+            DDimI3,
+            BcLower2,
+            BcUpper2,
+            BcLower3,
+            BcUpper3,
+            Solver>;
+#endif
 
     /// @brief The type of the first interpolation continuous dimension.
     using continuous_dimension_type1 = typename builder_type1::continuous_dimension_type;
@@ -224,6 +242,69 @@ public:
             deriv_type3>;
 
     /**
+     * @brief The type of the whole Derivs domain (cartesian product of the 2D Deriv domain
+     * and the associated batch domain) in the first and second dimensions, preserving the order
+     * of dimensions.
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y and Z,
+     * this is DiscreteDomain<Deriv<X>, Deriv<Y>, Z, T>.
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
+    using batched_derivs_domain_type1_2 = ddc::replace_dim_of_t<
+            ddc::replace_dim_of_t<
+                    BatchedInterpolationDDom,
+                    interpolation_discrete_dimension_type1,
+                    deriv_type1>,
+            interpolation_domain_type2,
+            deriv_type2>;
+
+    /**
+     * @brief The type of the whole Derivs domain (cartesian product of the 2D Deriv domain
+     * and the associated batch domain) in the second and third dimensions, preserving the order
+     * of dimensions.
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y, and Z
+     * this is DiscreteDomain<X, Deriv<Y>, Deriv<Z>, T>.
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
+    using batched_derivs_domain_type2_3 = ddc::replace_dim_of_t<
+            ddc::replace_dim_of_t<
+                    BatchedInterpolationDDom,
+                    interpolation_discrete_dimension_type2,
+                    deriv_type2>,
+            interpolation_domain_type3,
+            deriv_type3>;
+
+    /**
+     * @brief The type of the whole Derivs domain (cartesian product of the 2D Deriv domain
+     * and the associated batch domain) in the first and third dimensions, preserving the order
+     * of dimensions.
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
[…]
