### 2025-10-03 280daa1
Remove the flat 3d builder
