+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -14,8 +14,6 @@
 #include "spline_builder.hpp"
 #include "spline_builder_2d.hpp"
 
-#define USE_RECURSIVE_BUILDER
-
 namespace ddc {
 
 /**
@@ -55,23 +53,7 @@ public:
     using builder_type1 = ddc::
             SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
 
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate along second dimension.
-    using builder_type2 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate along third dimension.
-    using builder_type3 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower3, BcUpper3, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate the second-dimension-derivatives along first dimension.
-    using builder_deriv_type1 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate the third-dimension-derivatives along second dimension.
-    using builder_deriv_type2 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
-
-#ifdef USE_RECURSIVE_BUILDER
+    /// @brief The type of SplineBuilder used by this class to spline-approximate along the second and third dimensions.
     using builder_type2d = ddc::SplineBuilder2D<
             ExecSpace,
             MemorySpace,
@@ -84,16 +66,15 @@ public:
             BcLower3,
             BcUpper3,
             Solver>;
-#endif
 
     /// @brief The type of the first interpolation continuous dimension.
     using continuous_dimension_type1 = typename builder_type1::continuous_dimension_type;
 
     /// @brief The type of the second interpolation continuous dimension.
-    using continuous_dimension_type2 = typename builder_type2::continuous_dimension_type;
+    using continuous_dimension_type2 = typename builder_type2d::continuous_dimension_type1;
 
     /// @brief The type of the third interpolation continuous dimension.
-    using continuous_dimension_type3 = typename builder_type3::continuous_dimension_type;
+    using continuous_dimension_type3 = typename builder_type2d::continuous_dimension_type2;
 
     /// @brief The type of the first interpolation discrete dimension.
     using interpolation_discrete_dimension_type1 =
@@ -101,29 +82,29 @@ public:
 
     /// @brief The type of the second interpolation discrete dimension.
     using interpolation_discrete_dimension_type2 =
-            typename builder_type2::interpolation_discrete_dimension_type;
+            typename builder_type2d::interpolation_discrete_dimension_type1;
 
     /// @brief The type of the third interpolation discrete dimension.
     using interpolation_discrete_dimension_type3 =
-            typename builder_type3::interpolation_discrete_dimension_type;
+            typename builder_type2d::interpolation_discrete_dimension_type2;
 
     /// @brief The type of the B-splines in the first dimension.
     using bsplines_type1 = typename builder_type1::bsplines_type;
 
     /// @brief The type of the B-splines in the second dimension.
[…]
