### 2025-10-03 1b63b40
Use annotate_only in action-junit-report (#928)
