+++ .github/workflows/early_integration.yaml
@@ -78,8 +78,6 @@ jobs:
             cxx_version: '23'
     runs-on: ubuntu-latest
     needs: [docker-build]
-    permissions:
-      checks: write  # Required by mikepenz/action-junit-report
     steps:
       - name: Free Disk Space (Ubuntu)
         uses: jlumbroso/free-disk-space@54081f138730dfa15788a46383842cd2f914a1be  # v1.3.1
@@ -223,4 +221,5 @@ jobs:
         uses: mikepenz/action-junit-report@3585e9575db828022551b4231f165eb59a0e74e3  # v5.6.2
         if: ( success() || failure() ) && steps.test.outputs.with_report == 'true'  # always run even if the previous step fails
         with:
+          annotate_only: true  # forked repo cannot write to checks so just do annotations
           report_paths: '${{github.workspace}}/tests.xml'
