### 2025-10-05 5ffd14a
Simplify cmake targets exporting
