### 2025-10-06 dc1b812
Reorder the builder3d arguments to match the 2D builder
