+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -513,6 +513,30 @@ public:
      *      The values of the derivatives at the lower boundary in the third dimension.
      * @param[in] derivs_max3
      *      The values of the derivatives at the upper boundary in the third dimension.
+     * @param[in] mixed_derivs_min1_min2
+     *      The values of the the cross-derivatives at the lower boundary in the first dimension and the lower boundary in the second dimension.
+     * @param[in] mixed_derivs_max1_min2
+     *      The values of the the cross-derivatives at the upper boundary in the first dimension and the lower boundary in the second dimension.
+     * @param[in] mixed_derivs_min1_max2
+     *      The values of the the cross-derivatives at the lower boundary in the first dimension and the upper boundary in the second dimension.
+     * @param[in] mixed_derivs_max1_max2
+     *      The values of the the cross-derivatives at the upper boundary in the first dimension and the upper boundary in the second dimension.
+     * @param[in] mixed_derivs_min2_min3
+     *      The values of the the cross-derivatives at the lower boundary in the second dimension and the lower boundary in the third dimension.
+     * @param[in] mixed_derivs_max2_min3
+     *      The values of the the cross-derivatives at the upper boundary in the second dimension and the lower boundary in the third dimension.
+     * @param[in] mixed_derivs_min2_max3
+     *      The values of the the cross-derivatives at the lower boundary in the second dimension and the upper boundary in the third dimension.
+     * @param[in] mixed_derivs_max2_max3
+     *      The values of the the cross-derivatives at the upper boundary in the second dimension and the upper boundary in the third dimension.
+     * @param[in] mixed_derivs_min1_min3
+     *      The values of the the cross-derivatives at the lower boundary in the first dimension and the lower boundary in the third dimension.
+     * @param[in] mixed_derivs_max1_min3
+     *      The values of the the cross-derivatives at the upper boundary in the first dimension and the lower boundary in the third dimension.
+     * @param[in] mixed_derivs_min1_max3
+     *      The values of the the cross-derivatives at the lower boundary in the first dimension and the upper boundary in the third dimension.
+     * @param[in] mixed_derivs_max1_max3
+     *      The values of the the cross-derivatives at the upper boundary in the first dimension and the upper boundary in the third dimension.
      * @param[in] mixed_derivs_min1_min2_min3
      *      The values of the the cross-derivatives at the lower boundary in the first dimension, the lower boundary in the second dimension and the lower boundary in the third dimension.
      * @param[in] mixed_derivs_max1_min2_min3
@@ -584,13 +608,13 @@ public:
                     double const,
                     batched_derivs_domain_type1_2<BatchedInterpolationDDom>,
                     Layout,
-                    memory_space>> mixed_derivs_min1_max2
+                    memory_space>> mixed_derivs_max1_min2
             = std::nullopt,
             std::optional<ddc::ChunkSpan<
                     double const,
                     batched_derivs_domain_type1_2<BatchedInterpolationDDom>,
                     Layout,
-                    memory_space>> mixed_derivs_max1_min2
+                    memory_space>> mixed_derivs_min1_max2
             = std::nullopt,
             std::optional<ddc::ChunkSpan<
                     double const,
@@ -608,13 +632,13 @@ public:
                     double const,
                     batched_derivs_domain_type2_3<BatchedInterpolationDDom>,
                     Layout,
-                    memory_space>> mixed_derivs_min2_max3
+                    memory_space>> mixed_derivs_max2_min3
             = std::nullopt,
             std::optional<ddc::ChunkSpan<
[…]
