+++ tests/splines/batched_3d_spline_builder.cpp
@@ -983,24 +983,24 @@ void Batched3dSplineTest()
             std::optional(derivs3_lhs.span_cview()),
             std::optional(derivs3_rhs.span_cview()),
             std::optional(derivs_mixed_lhs1_lhs2.span_cview()),
-            std::optional(derivs_mixed_lhs1_rhs2.span_cview()),
             std::optional(derivs_mixed_rhs1_lhs2.span_cview()),
+            std::optional(derivs_mixed_lhs1_rhs2.span_cview()),
             std::optional(derivs_mixed_rhs1_rhs2.span_cview()),
             std::optional(derivs_mixed_lhs2_lhs3.span_cview()),
-            std::optional(derivs_mixed_lhs2_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs2_lhs3.span_cview()),
+            std::optional(derivs_mixed_lhs2_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs2_rhs3.span_cview()),
             std::optional(derivs_mixed_lhs1_lhs3.span_cview()),
-            std::optional(derivs_mixed_lhs1_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_lhs3.span_cview()),
+            std::optional(derivs_mixed_lhs1_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_rhs3.span_cview()),
             std::optional(derivs_mixed_lhs1_lhs2_lhs3.span_cview()),
-            std::optional(derivs_mixed_lhs1_rhs2_lhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_lhs2_lhs3.span_cview()),
+            std::optional(derivs_mixed_lhs1_rhs2_lhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_rhs2_lhs3.span_cview()),
             std::optional(derivs_mixed_lhs1_lhs2_rhs3.span_cview()),
-            std::optional(derivs_mixed_lhs1_rhs2_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_lhs2_rhs3.span_cview()),
+            std::optional(derivs_mixed_lhs1_rhs2_rhs3.span_cview()),
             std::optional(derivs_mixed_rhs1_rhs2_rhs3.span_cview()));
 #else
     spline_builder(coef, vals.span_cview());
