### 2025-10-07 edd99b4
Rename DDCTargets into DDCCoreTargets on @science-enthusiast's suggestion
