### 2025-10-07 a605c52
Update headers and forward declarations
