+++ include/ddc/chunk_common.hpp
@@ -4,17 +4,17 @@
 
 #pragma once
 
-#include <array>
 #include <cstddef>
 #include <type_traits>
 #include <utility>
 
 #include <Kokkos_Core.hpp>
 
-#include "detail/macros.hpp"
+#include "detail/type_seq.hpp"
 
 #include "chunk_traits.hpp"
 #include "discrete_domain.hpp"
+#include "discrete_vector.hpp"
 
 namespace ddc {
 
