+++ include/ddc/kernels/splines/greville_interpolation_points.hpp
@@ -10,6 +10,8 @@
 
 #include <ddc/ddc.hpp>
 
+#include <Kokkos_Core.hpp>
+
 #include "bsplines_non_uniform.hpp"
 #include "bsplines_uniform.hpp"
 #include "spline_boundary_conditions.hpp"
