+++ tests/sparse_discrete_domain.cpp
@@ -6,6 +6,9 @@
 
 #include <gtest/gtest.h>
 
+#include <Kokkos_Core.hpp>
+#include <Kokkos_StdAlgorithms.hpp>
+
 inline namespace anonymous_namespace_workaround_sparse_discrete_domain_cpp {
 
 using DElem0D = ddc::DiscreteElement<>;
