+++ include/ddc/kernels/fft.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cassert>
 #include <stdexcept>
 #include <type_traits>
 #include <utility>
