+++ include/ddc/kernels/splines/spline_builder_2d.hpp
@@ -11,6 +11,7 @@
 
 #include <ddc/ddc.hpp>
 
+#include "spline_boundary_conditions.hpp"
 #include "spline_builder.hpp"
 
 namespace ddc {
