+++ include/ddc/parallel_transform_reduce.hpp
@@ -11,10 +11,7 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "detail/kokkos.hpp"
-
 #include "ddc_to_kokkos_execution_policy.hpp"
-#include "discrete_domain.hpp"
 #include "discrete_element.hpp"
 #include "reducer.hpp"
 
