### 2025-10-07 58ed8e5
Update headers and forward declarations
