+++ tests/sparse_discrete_domain.cpp
@@ -6,6 +6,8 @@
 
 #include <gtest/gtest.h>
 
+#include <Kokkos_Core.hpp>
+
 inline namespace anonymous_namespace_workaround_sparse_discrete_domain_cpp {
 
 struct DDimX
