+++ tests/discrete_space.cpp
@@ -10,7 +10,9 @@
 
 inline namespace anonymous_namespace_workaround_discrete_space_cpp {
 
-struct DimX;
+struct DimX
+{
+};
 struct DDimX : ddc::UniformPointSampling<DimX>
 {
 };
