### 2025-10-08 182b5db
Use a StridedDiscreteDomain for the 1D builder's derivatives
