+++ include/ddc/kernels/splines/spline_builder.hpp
@@ -25,6 +25,27 @@
 
 namespace ddc {
 
+namespace detail {
+template <typename... DDims>
+struct make_discrete_vector
+{
+    template <std::size_t... Ints>
+    static ddc::DiscreteVector<DDims...> exec(std::index_sequence<Ints...>)
+    {
+        // Using the comma operator (Ints, 1)... gives an unused value warning
+        return ddc::DiscreteVector<DDims...>((Ints - Ints + 1)...);
+    }
+};
+template <typename... DDim>
+ddc::StridedDiscreteDomain<DDim...> to_strided_ddom(ddc::DiscreteDomain<DDim...> const& ddom)
+{
+    return ddc::StridedDiscreteDomain<DDim...>(
+            ddom.front(),
+            ddom.extents(),
+            make_discrete_vector<DDim...>::exec(std::make_index_sequence<sizeof...(DDim)> {}));
+}
+} // namespace detail
+
 /**
  * @brief An enum determining the backend solver of a SplineBuilder or SplineBuilder2d.
  *
@@ -153,6 +174,7 @@ private:
 
 public:
     /**
+     * TODO: update documentation
      * @brief The type of the whole Deriv domain (cartesian product of 1D Deriv domain
      * and batch domain) preserving the underlying memory layout (order of dimensions).
      *
@@ -164,10 +186,26 @@ public:
     template <
             class BatchedInterpolationDDom,
             class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    using batched_derivs_domain_type = ddc::replace_dim_of_t<
-            BatchedInterpolationDDom,
-            interpolation_discrete_dimension_type,
-            deriv_type>;
+    using whole_derivs_domain_type
+            = ddc::detail::convert_type_seq_to_strided_discrete_domain_t<ddc::type_seq_cat_t<
+                    ddc::detail::TypeSeq<deriv_type>,
+                    ddc::to_type_seq_t<BatchedInterpolationDDom>>>;
+
+    /**
+     * @brief The type of the whole Deriv domain (cartesian product of 1D Deriv domain
+     * and batch domain) preserving the underlying memory layout (order of dimensions).
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z> and a dimension of interest Y,
+     * this is DiscreteDomain<X,Deriv<Y>,Z>
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
+    using batched_derivs_domain_type = ddc::remove_dims_of_t<
+            whole_derivs_domain_type<BatchedInterpolationDDom>,
+            interpolation_discrete_dimension_type>;
 
     /// @brief Indicates if the degree of the splines is odd or even.
     static constexpr bool s_odd = BSplines::degree() % 2;
@@ -421,11 +459,13 @@ public:
             BatchedInterpolationDDom const& batched_interpolation_domain) const noexcept
     {
         assert(interpolation_domain() == interpolation_domain_type(batched_interpolation_domain));
-        return ddc::replace_dim_of<interpolation_discrete_dimension_type, deriv_type>(
-                batched_interpolation_domain,
-                ddc::DiscreteDomain<deriv_type>(
-                        ddc::DiscreteElement<deriv_type>(1),
-                        ddc::DiscreteVector<deriv_type>(s_nbc_xmin)));
+        return ddc::StridedDiscreteDomain(
+                detail::to_strided_ddom(
[…]
