### 2025-10-08 8778db4
Use a StridedDiscreteDomain for the 1D builder's derivatives
