### 2025-10-08 fc6a29e
Simplify cmake targets exporting (#930)
* Simplify cmake targets exporting

* Rename DDCTargets into DDCCoreTargets on @science-enthusiast's suggestion
