### 2025-10-08 27ed754
Allow print to work on any set
