### 2025-10-09 a1a9507
Reorder the dimensions of `whols_serivs_domain_type`
