+++ include/ddc/kernels/splines/spline_builder.hpp
@@ -17,6 +17,8 @@
 
 #include <Kokkos_Core.hpp>
 
+#include "ddc/detail/type_seq.hpp"
+
 #include "deriv.hpp"
 #include "integrals.hpp"
 #include "math_tools.hpp"
@@ -44,6 +46,63 @@ ddc::StridedDiscreteDomain<DDim...> to_strided_ddom(ddc::DiscreteDomain<DDim...>
             ddom.extents(),
             make_discrete_vector<DDim...>::exec(std::make_index_sequence<sizeof...(DDim)> {}));
 }
+
+// TODO: document, rename
+template <typename... T>
+struct to_whole_derivs_domain;
+
+template <typename... DDimsI, typename... DerivDims, typename... DDims>
+struct to_whole_derivs_domain<TypeSeq<DDimsI...>, TypeSeq<DerivDims...>, TypeSeq<DDims...>>
+{
+    using type = convert_type_seq_to_strided_discrete_domain_t<type_seq_cat_t<
+            TypeSeq<DDimsI...>,
+            type_seq_replace_t<TypeSeq<DDims...>, TypeSeq<DDimsI...>, TypeSeq<DerivDims...>>>>;
+};
+
+template <typename SeqDDimsI, typename SeqDerivDims, typename SeqDDims>
+using to_whole_derivs_domain_t =
+        typename to_whole_derivs_domain<SeqDDimsI, SeqDerivDims, SeqDDims>::type;
+
+template <
+        typename... DerivDims,
+        typename... DDimsI,
+        template <typename...> class DDomInterp,
+        typename... DDims,
+        template <typename...> class DDomBatched>
+auto get_whole_derivs_domain(
+        DDomInterp<DDimsI...> const& interpolation_dom,
+        DDomBatched<DDims...> const& batched_domain,
+        int degree)
+{
+    using result_domain_t = to_whole_derivs_domain_t<
+            TypeSeq<DDimsI...>,
+            TypeSeq<DerivDims...>,
+            TypeSeq<DDims...>>;
+    using DElem = typename result_domain_t::discrete_element_type;
+    using DVect = typename result_domain_t::discrete_vector_type;
+
+    auto const batch_domain = ddc::remove_dims_of<DDimsI...>(batched_domain);
+
+    return result_domain_t(
+            DElem(interpolation_dom.front(),
+                  batch_domain.front(),
+                  (ddc::DiscreteElement<DerivDims>(1))...),
+            DVect((ddc::DiscreteVector<DDimsI>(2))...,
+                  batch_domain.extents(),
+                  (ddc::DiscreteVector<DerivDims>(degree / 2))...),
+            DVect((ddc::DiscreteVector<DDimsI>(
+                          interpolation_dom.template extent<DDimsI>().value() - 2))...,
+                  typename decltype(batch_domain)::discrete_vector_type(
+                          (ddc::DiscreteVector<DDims>(1))...),
+                  (ddc::DiscreteVector<DerivDims>(1))...));
+}
+
+template <typename... DerivDims, typename... DDimsI, template <typename...> class DDomInterp>
+auto get_whole_derivs_domain(DDomInterp<DDimsI...> const& interpolation_dom, int degree)
+{
+    return get_whole_derivs_domain<DerivDims...>(interpolation_dom, interpolation_dom, degree);
+}
+
 } // namespace detail
 
 /**
@@ -181,15 +240,18 @@ public:
      * @tparam The batched discrete domain on which the interpolation points are defined.
      *
      * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z> and a dimension of interest Y,
-     * this is DiscreteDomain<X,Deriv<Y>,Z>
+     * this is StridedDiscreteDomain<Y,X,Deriv<Y>,Z>
      */
     template <
[…]
