### 2025-10-10 deae66e
Add 3D Splines (with periodic or greville boundary conditions) (#925)
* Add a 3D spline builder for the periodic case

* Changes to the derivs domain
- rename some of the type aliases
- use only one deriv dimension in the derivs_domain type aliases
- enable the greville tests
- lower the norm threshold for the triple derivative to 1e-10

* Reword the FIXMEs in the 3D test

* Fix a typo

* Format the splines CMakeLists.txt

* Add missing doc comments and add unused annotations to the derivs
parameters

* Declare new header files to CMake

* Increase timeout

* Modify 3D error bound computation

[…]
