+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -1,0 +1,708 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include <cassert>
+#include <cstddef>
+#include <optional>
+#include <type_traits>
+
+#include <ddc/ddc.hpp>
+
+#include "spline_builder.hpp"
+
+namespace ddc {
+
+/**
+ * @brief A class for creating a 3D spline approximation of a function.
+ *
+ * A class which contains an operator () which can be used to build a 3D spline approximation
+ * of a function. A 3D spline approximation uses a cross-product between three 1D SplineBuilder.
+ *
+ * @see SplineBuilder
+ */
+template <
+        class ExecSpace,
+        class MemorySpace,
+        class BSpline1,
+        class BSpline2,
+        class BSpline3,
+        class DDimI1,
+        class DDimI2,
+        class DDimI3,
+        ddc::BoundCond BcLower1,
+        ddc::BoundCond BcUpper1,
+        ddc::BoundCond BcLower2,
+        ddc::BoundCond BcUpper2,
+        ddc::BoundCond BcLower3,
+        ddc::BoundCond BcUpper3,
+        ddc::SplineSolver Solver>
+class SplineBuilder3D
+{
+public:
+    /// @brief The type of the Kokkos execution space used by this class.
+    using exec_space = ExecSpace;
+
+    /// @brief The type of the Kokkos memory space used by this class.
+    using memory_space = MemorySpace;
+
+    /// @brief The type of the SplineBuilder used by this class to spline-approximate along first dimension.
+    using builder_type1 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
+
+    /// @brief The type of the SplineBuilder used by this class to spline-approximate along second dimension.
+    using builder_type2 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
+
+    /// @brief The type of the SplineBuilder used by this class to spline-approximate along third dimension.
+    using builder_type3 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower3, BcUpper3, Solver>;
+
+    /// @brief The type of the SplineBuilder used by this class to spline-approximate the second-dimension-derivatives along first dimension.
+    using builder_deriv_type1 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
+
+    /// @brief The type of the SplineBuilder used by this class to spline-approximate the third-dimension-derivatives along second dimension.
+    using builder_deriv_type2 = ddc::
+            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
+
+    /// @brief The type of the first interpolation continuous dimension.
+    using continuous_dimension_type1 = typename builder_type1::continuous_dimension_type;
+
+    /// @brief The type of the second interpolation continuous dimension.
+    using continuous_dimension_type2 = typename builder_type2::continuous_dimension_type;
+
+    /// @brief The type of the third interpolation continuous dimension.
+    using continuous_dimension_type3 = typename builder_type3::continuous_dimension_type;
+
+    /// @brief The type of the first interpolation discrete dimension.
+    using interpolation_discrete_dimension_type1 =
+            typename builder_type1::interpolation_discrete_dimension_type;
+
+    /// @brief The type of the second interpolation discrete dimension.
[…]
