+++ include/ddc/kernels/splines.hpp
@@ -18,8 +18,10 @@
 #include "splines/spline_boundary_conditions.hpp"
 #include "splines/spline_builder.hpp"
 #include "splines/spline_builder_2d.hpp"
+#include "splines/spline_builder_3d.hpp"
 #include "splines/spline_evaluator.hpp"
 #include "splines/spline_evaluator_2d.hpp"
+#include "splines/spline_evaluator_3d.hpp"
 #include "splines/spline_traits.hpp"
 #include "splines/splines_linear_problem.hpp"
 #include "splines/splines_linear_problem_2x2_blocks.hpp"
