+++ CMakeLists.txt
@@ -282,8 +282,10 @@ if("${DDC_BUILD_KERNELS_SPLINES}")
                     include/ddc/kernels/splines/spline_boundary_conditions.hpp
                     include/ddc/kernels/splines/spline_builder.hpp
                     include/ddc/kernels/splines/spline_builder_2d.hpp
+                    include/ddc/kernels/splines/spline_builder_3d.hpp
                     include/ddc/kernels/splines/spline_evaluator.hpp
                     include/ddc/kernels/splines/spline_evaluator_2d.hpp
+                    include/ddc/kernels/splines/spline_evaluator_3d.hpp
                     include/ddc/kernels/splines/spline_traits.hpp
                     include/ddc/kernels/splines/splines_linear_problem.hpp
                     include/ddc/kernels/splines/splines_linear_problem_2x2_blocks.hpp

[…]
