+++ search/all_c.js
@@ -1,12 +1,12 @@
 var searchData=
 [
   ['ifft_0',['ifft',['../namespaceddc.html#a17507ff514a63591f7e9e4822e1ebacc',1,'ddc']]],
-  ['impl_1',['impl',['../classddc_1_1PeriodicSampling_1_1Impl.html#adcaf2527acc2c7a0a18f741240082b56',1,'ddc::PeriodicSampling::Impl::Impl()'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#a641f57a88c58bb99afa808bf60899b8b',1,'ddc::NonUniformPointSampling::Impl::Impl(InputIt const points_begin, InputIt const points_end)'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#aa31034a9da59ee6eecbd14cca6e31dae',1,'ddc::NonUniformPointSampling::Impl::Impl(InputRange const &amp;points)'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#ae2a5749211782b11c07db0c5e18aabeb',1,'ddc::NonUniformPointSampling::Impl::Impl(std::initializer_list&lt; Coordinate&lt; CDim &gt; &gt; const points)'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#a303cc88f434dd361877b568c23f93203',1,'ddc::NonUniformPointSampling::Impl::Impl()=default'],['../classddc_1_1UniformBSplines_1_1Impl.html#a72335eea98a83b669ed3eadb4ea52cf4',1,'ddc::UniformBSplines::Impl::Impl(Impl &amp;&amp;x)=default'],['../classddc_1_1UniformBSplines_1_1Impl.html#a669cca92a0ef2eb8ad79668c59bacaa5',1,'ddc::UniformBSplines::Impl::Impl(Impl const &amp;x)=default'],['../classddc_1_1UniformBSplines_1_1Impl.html#aad14ec609c7f8fa7a92559cc9c442d7e',1,'ddc::UniformBSplines::Impl::Impl(Impl&lt; DDim, OriginMemorySpace &gt; const &amp;impl)'],['../classddc_1_1UniformBSplines_1_1Impl.html#a1027ba57ec532f30acea8614c4b016bd',1,'ddc::UniformBSplines::Impl::Impl(ddc::Coordinate&lt; CDim &gt; rmin, ddc::Coordinate&lt; CDim &gt; rmax, std::size_t ncells)'],['../classddc_1_1NonUniformBSplines_1_1Impl.html#adcaf2527acc2c7a0a18f741240082b56',1,'ddc::NonUniformBSplines::Impl::Impl()'],['../classddc_1_1UniformBSplines_1_1Impl.html#adcaf2527acc2c7a0a18f741240082b56',1,'ddc::UniformBSplines::Impl::Impl()'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#adcaf2527acc2c7a0a18f741240082b56',1,'ddc::NonUniformPointSampling::Impl::Impl'],['../classddc_1_1NonUniformPointSampling_1_1Impl.html#a290dbdbd06bfcf85cc6f050fdce7ec60',1,'ddc::NonUniformPointSampling::Impl::Impl(Impl&lt; DDim, OriginMemorySpace &gt; const &amp;impl)'],['../classddc_1_1UniformPointSampling_1_1Impl.html#adcaf2527acc2c7a0a18f741240082b56',1,'ddc::UniformPointSampling::Impl::Impl()'],['../classddc_1_1NonUniformBSplines_1_1Impl.html#ab6ddf1cfc130d5fd3d67e36d1cab51a6',1,'ddc::NonUniformBSplines::Impl::Impl(
[…]
