+++ search/functions_13.js
@@ -2,6 +2,7 @@ var searchData=
 [
   ['uid_0',['uid',['../classddc_1_1DiscreteElement.html#a1cf25c7978e4e81cb194966f7a7ae154',1,'ddc::DiscreteElement::uid() noexcept'],['../classddc_1_1DiscreteElement.html#a137ec8ab82fce4d40d7f763c0f277586',1,'ddc::DiscreteElement::uid() const noexcept'],['../classddc_1_1DiscreteElement.html#ac40188cb36135f8260b73696f3ee7242',1,'ddc::DiscreteElement::uid() noexcept'],['../classddc_1_1DiscreteElement.html#ae99e00a72ea441716e17a5eaa3a7c65f',1,'ddc::DiscreteElement::uid() const noexcept'],['../namespaceddc.html#a70f6a7634ef3f77720ba76b015de2008',1,'ddc::uid(DiscreteElement&lt; Tag &gt; const &amp;tuple) noexcept'],['../namespaceddc.html#abfeecbd05212f790cbdd85181f01c59e',1,'ddc::uid(DiscreteElement&lt; Tag &gt; &amp;tuple) noexcept'],['../namespaceddc.html#ad055e25ad4b426fad4ea6b650e000755',1,'ddc::uid(DiscreteElement&lt; Tags... &gt; const &amp;tuple) noexcept'],['../namespaceddc.html#a9a4cd62d6f8f234732cf71afa0859f09',1,'ddc::uid(DiscreteElement&lt; Tags... &gt; &amp;tuple) noexcept']]],
   ['upper_5fextrapolation_5frule_1',['upper_extrapolation_rule',['../classddc_1_1SplineEvaluator.html#a2955c7bfc010d401f09e2078a5a79544',1,'ddc::SplineEvaluator']]],
-  ['upper_5fextrapolation_5frule_5fdim_5f1_2',['upper_extrapolation_rule_dim_1',['../classddc_1_1SplineEvaluator2D.html#ad7e2ce4dba00ac5801987978cd42e169',1,'ddc::SplineEvaluator2D']]],
-  ['upper_5fextrapolation_5frule_5fdim_5f2_3',['upper_extrapolation_rule_dim_2',['../classddc_1_1SplineEvaluator2D.html#ad326988914e40f7e23a32396b6607d60',1,'ddc::SplineEvaluator2D']]]
+  ['upper_5fextrapolation_5frule_5fdim_5f1_2',['upper_extrapolation_rule_dim_1',['../classddc_1_1SplineEvaluator2D.html#ad7e2ce4dba00ac5801987978cd42e169',1,'ddc::SplineEvaluator2D::upper_extrapolation_rule_dim_1()'],['../classddc_1_1SplineEvaluator3D.html#a0aae7a18509c48a03957d0bda19b4e38',1,'ddc::SplineEvaluator3D::upper_extrapolation_rule_dim_1()']]],
+  ['upper_5fextrapolation_5frule_5fdim_5f2_3',['upper_extrapolation_rule_dim_2',['../classddc_1_1SplineEvaluator2D.html#ad326988914e40f7e23a32396b6607d60',1,'ddc::SplineEvaluator2D::upper_extrapolation_rule_dim_2()'],['../classddc_1_1SplineEvaluator3D.html#a65382d71efa995645ff6c9838869038e',1,'ddc::SplineEvaluator3D::upper_extrapolation_rule_dim_2() const']]],
+  ['upper_5fextrapolation_5frule_5fdim_5f3_4',['upper_extrapolation_rule_dim_3',['../classddc_1_1SplineEvaluator3D.html#a34acac0c85a62f94564e20bcd503c5a9',1,'ddc::SplineEvaluator3D']]]
 ];

[…]
