### 2025-10-10 0e6402e
Use StridedDiscreteDomain for the 2D builder's derivatives
