+++ include/ddc/kernels/splines/spline_builder_2d.hpp
@@ -15,6 +15,111 @@
 
 namespace ddc {
 
+namespace detail {
+template <typename ElementType, typename Layout, typename MemorySpace, typename... DDims>
+auto strided_to_discrete_domain_chunkspan(
+        ddc::ChunkSpan<
+                ElementType,
+                ddc::StridedDiscreteDomain<DDims...>,
+                Layout,
+                MemorySpace> const& cs)
+{
+    assert(((cs.domain().strides().template get<DDims>() == 1) && ...));
+    return ddc::ChunkSpan<ElementType, ddc::DiscreteDomain<DDims...>, Layout, MemorySpace>(
+            cs.data_handle(),
+            ddc::DiscreteDomain<DDims...>(cs.domain().front(), cs.domain().extents()));
+}
+
+template <typename ElementType, typename Layout, typename MemorySpace, typename... DDims>
+auto discrete_to_strided_domain_chunkspan(
+        ddc::ChunkSpan<ElementType, ddc::DiscreteDomain<DDims...>, Layout, MemorySpace> const& cs)
+{
+    return ddc::ChunkSpan<ElementType, ddc::StridedDiscreteDomain<DDims...>, Layout, MemorySpace>(
+            cs.data_handle(),
+            ddc::StridedDiscreteDomain<DDims...>(
+                    cs.domain().front(),
+                    cs.domain().extents(),
+                    make_discrete_vector<DDims...>(1)));
+}
+
+template <typename Dim>
+struct Min
+{
+};
+
+template <typename Dim>
+constexpr Min<Dim> dmin;
+
+template <typename Dim>
+struct Max
+{
+};
+
+template <typename Dim>
+constexpr Max<Dim> dmax;
+
+template <typename... DDims, typename T, typename Support, class Layout, class MemorySpace>
+auto derivs(
+        ddc::ChunkSpan<T, Support, Layout, MemorySpace> const& derivs_span,
+        ddc::DiscreteElement<DDims...> delem)
+{
+    return derivs_span[delem];
+}
+
+template <
+        typename DerivDim,
+        typename... DerivDims,
+        typename... DDims,
+        typename T,
+        typename Support,
+        class Layout,
+        class MemorySpace>
+auto derivs(
+        ddc::ChunkSpan<T, Support, Layout, MemorySpace> const& derivs_span,
+        ddc::DiscreteElement<DDims...> delem,
+        Min<DerivDim>,
+        DerivDims... deriv_dims)
+{
+    return derivs(
+            derivs_span,
+            ddc::DiscreteElement<
+                    DDims...,
+                    DerivDim>(delem, ddc::DiscreteElement<DerivDim>(derivs_span.domain().front())),
+            deriv_dims...);
+}
+
+template <
+        typename DerivDim,
+        typename... DerivDims,
+        typename... DDims,
+        typename T,
+        typename Support,
+        class Layout,
+        class MemorySpace>
+auto derivs(
+        ddc::ChunkSpan<T, Support, Layout, MemorySpace> const& derivs_span,
+        ddc::DiscreteElement<DDims...> delem,
+        Max<DerivDim>,
+        DerivDims... deriv_dims)
+{
+    return derivs(
+            derivs_span,
+            ddc::DiscreteElement<
+                    DDims...,
+                    DerivDim>(delem, ddc::DiscreteElement<DerivDim>(derivs_span.domain().back())),
+            deriv_dims...);
+}
+
+template <typename... DerivDims, typename T, typename Support, class Layout, class MemorySpace>
+auto derivs(
+        ddc::ChunkSpan<T, Support, Layout, MemorySpace> const& derivs_span,
+        DerivDims... deriv_dims)
+{
[…]
