### 2025-10-10 0c92739
Use StridedDiscreteDomain for the 2D builder's derivatives
