+++ tests/splines/batched_2d_spline_builder.cpp
@@ -4,9 +4,7 @@
 
 #include <algorithm>
 #include <cstddef>
-#if defined(BC_HERMITE)
-#    include <optional>
-#endif
+
 #if defined(BSPLINES_TYPE_UNIFORM)
 #    include <type_traits>
 #endif
@@ -184,21 +182,28 @@ void Batched2dSplineTest()
     ddc::DiscreteDomain<DDims...> const
             dom_vals(dom_vals_tmp, interpolation_domain1, interpolation_domain2);
 
-#if defined(BC_HERMITE)
     // Create the derivs domain
     ddc::DiscreteDomain<ddc::Deriv<I1>> const
-            derivs_domain1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
+            derivs_ddom1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I2>> const
-            derivs_domain2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
+            derivs_ddom2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const
-            derivs_domain(derivs_domain1, derivs_domain2);
+            derivs_ddom(derivs_ddom1, derivs_ddom2);
 
-    auto const dom_derivs_1d
-            = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_vals, derivs_domain1);
-    auto const dom_derivs2 = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_vals, derivs_domain2);
-    auto const dom_derivs
-            = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs_1d, derivs_domain2);
-#endif
+    ddc::StridedDiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const derivs_domain(
+            DElem<ddc::Deriv<I1>, ddc::Deriv<I2>>(1, 1),
+            DVect<ddc::Deriv<I1>, ddc::Deriv<I2>>(s_degree / 2, s_degree / 2),
+            DVect<ddc::Deriv<I1>, ddc::Deriv<I2>>(1, 1));
+
+    auto const whole_derivs_domain1 = ddc::detail::get_whole_derivs_domain<
+            ddc::Deriv<I1>>(interpolation_domain1, dom_vals, s_degree);
+    auto const whole_derivs_domain2 = ddc::detail::get_whole_derivs_domain<
+            ddc::Deriv<I2>>(interpolation_domain2, dom_vals, s_degree);
+    auto const whole_derivs_domain = ddc::detail::get_whole_derivs_domain<
+            ddc::Deriv<I1>,
+            ddc::Deriv<I2>>(interpolation_domain, dom_vals, s_degree);
+
+    auto const dom_derivs = ddc::remove_dims_of<DDimI1, DDimI2>(whole_derivs_domain);
 
     // Create a SplineBuilder over BSplines<I> and batched along other dimensions using some boundary conditions
     ddc::SplineBuilder2D<
@@ -236,141 +241,156 @@ void Batched2dSplineTest()
                 vals(e) = vals_1d(DElem<DDimI1, DDimI2>(e));
             });
 
-#if defined(BC_HERMITE)
     // Allocate and fill a chunk containing derivs to be passed as input to spline_builder.
     int const shift = s_degree % 2; // shift = 0 for even order, 1 for odd order
[…]
