+++ tests/splines/multiple_batch_domain_2d_spline_builder.cpp
@@ -5,9 +5,6 @@
 #include <algorithm>
 #include <cstddef>
 #include <tuple>
-#if defined(BC_HERMITE)
-#    include <optional>
-#endif
 #if defined(BSPLINES_TYPE_UNIFORM)
 #    include <type_traits>
 #endif
@@ -176,28 +173,37 @@ std::tuple<double, double, double, double> ComputeEvaluationError(
     using I1 = typename DDimI1::continuous_dimension_type;
     using I2 = typename DDimI2::continuous_dimension_type;
 
-#if defined(BC_HERMITE)
-    ddc::DiscreteDomain<DDimI1> const interpolation_domain1(dom_vals);
-    ddc::DiscreteDomain<DDimI2> const interpolation_domain2(dom_vals);
-    // Create the derivs domain
-    ddc::DiscreteDomain<ddc::Deriv<I1>> const
-            derivs_domain1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
-    ddc::DiscreteDomain<ddc::Deriv<I2>> const
-            derivs_domain2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
-    ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const
-            derivs_domain(derivs_domain1, derivs_domain2);
-
-    auto const dom_derivs_1d
-            = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_vals, derivs_domain1);
-    auto const dom_derivs2 = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_vals, derivs_domain2);
-    auto const dom_derivs
-            = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs_1d, derivs_domain2);
-#endif
-
-    // Compute useful domains (dom_interpolation, dom_batch, dom_bsplines and dom_spline)
+    // Compute useful domains
     ddc::DiscreteDomain<DDimI1, DDimI2> const dom_interpolation
             = spline_builder.interpolation_domain();
     auto const dom_spline = spline_builder.batched_spline_domain(dom_vals);
+
+    ddc::DiscreteDomain<DDimI1> const interpolation_domain1(dom_vals);
+    ddc::DiscreteDomain<DDimI2> const interpolation_domain2(dom_vals);
+    ddc::DiscreteDomain<DDimI1, DDimI2> const
+            interpolation_domain(interpolation_domain1, interpolation_domain2);
+
+    // Create the derivs domain
+    ddc::DiscreteDomain<ddc::Deriv<I1>> const
+            derivs_ddom1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
+    ddc::DiscreteDomain<ddc::Deriv<I2>> const
+            derivs_ddom2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
+    ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const
+            derivs_ddom(derivs_ddom1, derivs_ddom2);
+
+    ddc::StridedDiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>> const derivs_domain(
+            DElem<ddc::Deriv<I1>, ddc::Deriv<I2>>(1, 1),
+            DVect<ddc::Deriv<I1>, ddc::Deriv<I2>>(s_degree / 2, s_degree / 2),
[…]
