### 2025-10-11 5a1d3b9
Fix some bugs and modify the interface of `get_whole_derivs_domain` - Fix code that was wrong in some assertions - Modify `get_whole_derivs_domain` to take the number of constraints for   each dimension instead of the degree
