### 2025-10-11 fb49435
Add missing doc comments
