### 2025-10-12 9f5dd70
Define `batched_derivs_domain_type1` the same way as the other aliases
