+++ .github/workflows/early_integration.yaml
@@ -191,7 +191,7 @@ jobs:
             cmake --build build
             if [ 'xcpu' = 'x${{matrix.backend.name}}' ]
             then
-              ctest --test-dir build --output-on-failure --timeout 10 --output-junit tests.xml
+              ctest --test-dir build --output-on-failure --timeout 20 --output-junit tests.xml
               ./build/examples/characteristics_advection
               ./build/examples/game_of_life
               ./build/examples/heat_equation_spectral
