+++ include/ddc/parallel_for_each.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include "parallel_for_each.hpp"
