+++ include/ddc/sparse_discrete_domain.hpp
@@ -5,14 +5,15 @@
 #pragma once
 
 #include <cstddef>
+#include <functional>
 #include <iterator>
+#include <string>
 #include <tuple>
 #include <type_traits>
 
 #include <Kokkos_Core.hpp>
 #include <Kokkos_StdAlgorithms.hpp>
 
-#include "detail/kokkos.hpp"
 #include "detail/tagged_vector.hpp"
 #include "detail/type_seq.hpp"
 
@@ -21,11 +22,8 @@
 
 namespace ddc {
 
-template <class DDim>
-struct SparseDiscreteDomainIterator;
-
 template <class... DDims>
-class SparseDiscreteDomain;
+class SparseDiscreteDomain; // IWYU pragma: keep
 
 template <class T>
 struct is_sparse_discrete_domain : std::false_type
