+++ include/ddc/create_mirror.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include "create_mirror.hpp"
