+++ examples/strided_discrete_domain.cpp
@@ -1,0 +1,10 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include <ddc/strided_discrete_domain.hpp>
+
+int main()
+{
+    return 0;
+}
