+++ include/ddc/kernels/splines/splines_linear_problem_pds_band.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include "splines_linear_problem_pds_band.hpp"
