+++ include/ddc/uniform_point_sampling.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include "uniform_point_sampling.hpp"
