+++ tests/splines/multiple_batch_domain_2d_spline_builder.cpp
@@ -8,9 +8,7 @@
 #if defined(BC_HERMITE)
 #    include <optional>
 #endif
-#if defined(BSPLINES_TYPE_UNIFORM)
-#    include <type_traits>
-#endif
+#include <string>
 #include <vector>
 
 #include <ddc/ddc.hpp>
