+++ tests/splines/splines_linear_problem.cpp
@@ -13,7 +13,6 @@
 #include <utility>
 #include <vector>
 
-#include <ddc/ddc.hpp>
 #include <ddc/kernels/splines.hpp>
 
 #include <gtest/gtest.h>
