+++ tests/type_seq.cpp
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+#include <string>
 #include <type_traits>
 
 #include <ddc/ddc.hpp>
@@ -10,13 +11,27 @@
 
 inline namespace anonymous_namespace_workaround_type_seq_cpp {
 
-struct a;
-struct b;
-struct c;
-struct d;
-struct e;
-struct y;
-struct z;
+struct a
+{
+};
+struct b
+{
+};
+struct c
+{
+};
+struct d
+{
+};
+struct e
+{
+};
+struct y
+{
+};
+struct z
+{
+};
 
 } // namespace anonymous_namespace_workaround_type_seq_cpp
 
