+++ include/ddc/kernels/splines/kokkos-kernels-ext/KokkosBatched_Gbtrs.hpp
@@ -8,7 +8,7 @@
 #ifndef KOKKOSBATCHED_GBTRS_HPP_
 #define KOKKOSBATCHED_GBTRS_HPP_
 
-#include <KokkosBatched_Util.hpp>
+#include <Kokkos_Macros.hpp>
 
 /// \author Yuuichi Asahi (yuuichi.asahi@cea.fr)
 
@@ -51,7 +51,7 @@ struct SerialGbtrs {
 };
 }  // namespace KokkosBatched
 
-#include "KokkosBatched_Gbtrs_Serial_Impl.hpp"
+#include "KokkosBatched_Gbtrs_Serial_Impl.hpp" // IWYU pragma: export
 
 #endif  // KOKKOSBATCHED_GBTRS_HPP_
 
