+++ include/ddc/discrete_domain.hpp
@@ -8,7 +8,9 @@
 #include <iterator>
 #include <tuple>
 #include <type_traits>
+#include <utility>
 
+#include <Kokkos_Assert.hpp>
 #include <Kokkos_Macros.hpp>
 
 #include "detail/type_seq.hpp"
@@ -22,7 +24,7 @@ template <class DDim>
 struct DiscreteDomainIterator;
 
 template <class... DDims>
-class DiscreteDomain;
+class DiscreteDomain; // IWYU pragma: keep
 
 template <class T>
 struct is_discrete_domain : std::false_type
