+++ include/ddc/kernels/splines/null_extrapolation_rule.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include "null_extrapolation_rule.hpp"
