+++ tests/splines/knots_as_interpolation_points.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 #include <cstddef>
+#include <string>
 #include <tuple>
 #include <type_traits>
 #include <utility>
@@ -18,7 +19,9 @@
 inline namespace anonymous_namespace_workaround_knots_as_interpolation_points_cpp {
 
 template <class T>
-struct UniformBSplinesFixture;
+struct UniformBSplinesFixture
+{
+};
 
 template <bool IsPeriodic>
 struct UniformBSplinesFixture<std::tuple<std::integral_constant<bool, IsPeriodic>>>
@@ -42,7 +45,9 @@ struct UniformBSplinesFixture<std::tuple<std::integral_constant<bool, IsPeriodic
 };
 
 template <class T>
-struct NonUniformBSplinesFixture;
+struct NonUniformBSplinesFixture
+{
+};
 
 template <bool IsPeriodic>
 struct NonUniformBSplinesFixture<std::tuple<std::integral_constant<bool, IsPeriodic>>>
