+++ include/ddc/discrete_domain.hpp
@@ -9,6 +9,7 @@
 #include <iterator>
 #include <tuple>
 #include <type_traits>
+#include <utility>
 
 #include <Kokkos_Macros.hpp>
 
@@ -23,7 +24,7 @@ template <class DDim>
 struct DiscreteDomainIterator;
 
 template <class... DDims>
-class DiscreteDomain;
+class DiscreteDomain; // IWYU pragma: keep
 
 template <class T>
 struct is_discrete_domain : std::false_type
