+++ include/ddc/transform_reduce.hpp
@@ -7,6 +7,7 @@
 #include <utility>
 
 #include "detail/macros.hpp"
+#include "detail/type_seq.hpp"
 
 #include "discrete_domain.hpp"
 #include "discrete_element.hpp"
