+++ include/ddc/uniform_point_sampling.hpp
@@ -19,6 +19,8 @@
 #include "discrete_vector.hpp"
 #include "real_type.hpp"
 
+// IWYU pragma: no_forward_declare Kokkos::HostSpace
+
 namespace ddc {
 
 namespace detail {
