+++ include/ddc/kernels/splines/kokkos-kernels-ext/KokkosBatched_Gbtrs_Serial_Impl.cpp
@@ -1,0 +1,16 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+// clang-format off
+// NOLINTBEGIN(*)
+
+#ifndef KOKKOSBATCHED_GBTRS_SERIAL_IMPL_HPP_
+#define KOKKOSBATCHED_GBTRS_SERIAL_IMPL_HPP_
+
+#include "KokkosBatched_Gbtrs_Serial_Impl.hpp"
+
+#endif  // KOKKOSBATCHED_GBTRS_SERIAL_IMPL_HPP_
+
+// NOLINTEND(*)
+// clang-format on
