### 2025-10-13 d21669a
Bump fsfe/reuse-action from 5.0.0 to 6.0.0 (#941)
Bumps [fsfe/reuse-action](https://github.com/fsfe/reuse-action) from 5.0.0 to 6.0.0.
- [Release notes](https://github.com/fsfe/reuse-action/releases)
- [Commits](https://github.com/fsfe/reuse-action/compare/bb774aa972c2a89ff34781233d275075cbddf542...676e2d560c9a403aa252096d99fcab3e1132b0f5)

---
updated-dependencies:
- dependency-name: fsfe/reuse-action
  dependency-version: 6.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bo
[…]
