### 2025-10-13 ca17d55
Bump mikepenz/action-junit-report from 5.6.2 to 6.0.0 (#943)
Bumps [mikepenz/action-junit-report](https://github.com/mikepenz/action-junit-report) from 5.6.2 to 6.0.0.
- [Release notes](https://github.com/mikepenz/action-junit-report/releases)
- [Commits](https://github.com/mikepenz/action-junit-report/compare/3585e9575db828022551b4231f165eb59a0e74e3...5b7ee5a21e8674b695313d769f3cbdfd5d4d53a4)

---
updated-dependencies:
- dependency-name: mikepenz/action-junit-report
  dependency-version: 6.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: 
[…]
