+++ include/ddc/kernels/splines/spline_builder_2d.hpp
@@ -637,12 +637,6 @@ operator()(
     // TODO: perform computations along dimension 1 on different streams ?
 
     // Spline1-approximate derivs_min2 (to spline1_deriv_min)
-    whole_derivs_domain_type2<BatchedInterpolationDDom> const batched_deriv_domain
-            = ddc::detail::get_whole_derivs_domain<deriv_type2>(
-                    ddc::select<ddim2>(batched_interpolation_domain),
-                    batched_interpolation_domain,
-                    bsplines_type2::degree() / 2);
-
     auto const spline_batched_deriv_domain = ddc::detail::get_whole_derivs_domain<deriv_type2>(
             ddc::select<ddim2>(batched_interpolation_domain),
             m_spline_builder_deriv1.batched_spline_domain(batched_interpolation_domain),
