### 2025-10-13 405f7f0
Use StridedDiscreteDomain for the derivs in the 3D builder
