### 2025-10-13 420fb1a
Use StridedDiscreteDomain for the derivs in the 3D builder
