+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -191,6 +191,24 @@ public:
             deriv_type1>;
 
     /**
+     * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
+     * product of 1D Deriv domain and batch domain) in the first dimension, to be passed as
+     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y and Z,
+     * this is StridedDiscreteDomain<X, Deriv<X>, Y, Z, T>
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
+    using whole_derivs_domain_type1 = detail::to_whole_derivs_domain_t<
+            ddc::detail::TypeSeq<interpolation_discrete_dimension_type1>,
+            ddc::detail::TypeSeq<deriv_type1>,
+            ddc::to_type_seq_t<BatchedInterpolationDDom>>;
+
+    /**
      * @brief The type of the whole Derivs domain (cartesian product of the 1D Deriv domain
      * and the associated batch domain) in the second dimension, preserving the order of dimensions.
      *
@@ -208,6 +226,24 @@ public:
             deriv_type2>;
 
     /**
+     * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
+     * product of 1D Deriv domain and batch domain) in the second dimension, to be passed as
+     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y and Z,
+     * this is StridedDiscreteDomain<Y, X, Deriv<Y>, Z, T>
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
+    using whole_derivs_domain_type2 = detail::to_whole_derivs_domain_t<
+            ddc::detail::TypeSeq<interpolation_discrete_dimension_type2>,
+            ddc::detail::TypeSeq<deriv_type2>,
+            ddc::to_type_seq_t<BatchedInterpolationDDom>>;
+
+    /**
      * @brief The type of the whole Derivs domain (cartesian product of the 1D Deriv domain
      * and the associated batch domain) in the third dimension, preserving the order of dimensions.
      *
@@ -223,6 +259,24 @@ public:
             BatchedInterpolationDDom,
             interpolation_discrete_dimension_type3,
             deriv_type3>;
+
+    /**
+     * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
+     * product of 1D Deriv domain and batch domain) in the third dimension, to be passed as
+     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     *
+     * @tparam The batched discrete domain on which the interpolation points are defined.
+     *
+     * Example: For batched_interpolation_domain_type = DiscreteDomain<X,Y,Z,T> and dimensions of interest X, Y and Z,
+     * this is StridedDiscreteDomain<Z, X, Y, Deriv<Z>, T>
+     */
+    template <
+            class BatchedInterpolationDDom,
+            class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
[…]
