+++ tests/splines/batched_3d_spline_builder.cpp
@@ -223,7 +223,7 @@ void Batched3dSplineTest()
     ddc::DiscreteDomain<ddc::Deriv<I1>, DDimI2, ddc::Deriv<I3>> const
             derivs_domain13(derivs_domain1, interpolation_domain2, derivs_domain3);
     ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>, ddc::Deriv<I3>> const
-            derivs_domain(derivs_domain1, derivs_domain2, derivs_domain3);
+            derivs_domain_all(derivs_domain1, derivs_domain2, derivs_domain3);
 
     auto const dom_derivs1 = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_vals, derivs_domain1);
     auto const dom_derivs2 = ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_vals, derivs_domain2);
@@ -234,9 +234,8 @@ void Batched3dSplineTest()
             = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs2, derivs_domain3);
     auto const dom_derivs13
             = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs1, derivs_domain3);
-    auto const dom_derivs = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(
-            ddc::replace_dim_of<DDimI2, ddc::Deriv<I2>>(dom_derivs1, derivs_domain2),
-            derivs_domain3);
+    auto const dom_derivs_all
+            = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs12, derivs_domain3);
 #endif
 
     // Create a SplineBuilder over BSplines<I> and batched along other dimensions using some boundary conditions
@@ -280,6 +279,7 @@ void Batched3dSplineTest()
             });
 
 #if defined(BC_HERMITE)
+    // Allocate and fill a chunk containing derivs to be passed as input to spline_builder.
     int const shift = s_degree % 2; // shift = 0 for even order, 1 for odd order
 
     ddc::Chunk derivs1_lhs_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
@@ -751,92 +751,102 @@ void Batched3dSplineTest()
     }
 
     ddc::Chunk derivs_mixed_lhs1_lhs2_lhs3_alloc(
-            dom_derivs,
+            dom_derivs_all,
             ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_lhs1_lhs2_lhs3
             = derivs_mixed_lhs1_lhs2_lhs3_alloc.span_view();
     ddc::Chunk derivs_mixed_rhs1_lhs2_lhs3_alloc(
-            dom_derivs,
+            dom_derivs_all,
             ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_rhs1_lhs2_lhs3
             = derivs_mixed_rhs1_lhs2_lhs3_alloc.span_view();
     ddc::Chunk derivs_mixed_lhs1_rhs2_lhs3_alloc(
-            dom_derivs,
+            dom_derivs_all,
             ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_lhs1_rhs2_lhs3
             = derivs_mixed_lhs1_rhs2_lhs3_alloc.span_view();
     ddc::Chunk derivs_mixed_rhs1_rhs2_lhs3_alloc(
-            dom_derivs,
+            dom_derivs_all,
[…]
