### 2025-10-14 d2835bf
Reorder some of the deris cunkspan declarations to match the 2D tests
