+++ tests/splines/batched_3d_spline_builder.cpp
@@ -478,11 +478,11 @@ void Batched3dSplineTest()
             derivs_mixed_lhs1_lhs2_alloc(dom_derivs12, ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_lhs1_lhs2 = derivs_mixed_lhs1_lhs2_alloc.span_view();
     ddc::Chunk
-            derivs_mixed_lhs1_rhs2_alloc(dom_derivs12, ddc::KokkosAllocator<double, MemorySpace>());
-    ddc::ChunkSpan const derivs_mixed_lhs1_rhs2 = derivs_mixed_lhs1_rhs2_alloc.span_view();
-    ddc::Chunk
             derivs_mixed_rhs1_lhs2_alloc(dom_derivs12, ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_rhs1_lhs2 = derivs_mixed_rhs1_lhs2_alloc.span_view();
+    ddc::Chunk
+            derivs_mixed_lhs1_rhs2_alloc(dom_derivs12, ddc::KokkosAllocator<double, MemorySpace>());
+    ddc::ChunkSpan const derivs_mixed_lhs1_rhs2 = derivs_mixed_lhs1_rhs2_alloc.span_view();
     ddc::Chunk
             derivs_mixed_rhs1_rhs2_alloc(dom_derivs12, ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs_mixed_rhs1_rhs2 = derivs_mixed_rhs1_rhs2_alloc.span_view();
@@ -491,12 +491,12 @@ void Batched3dSplineTest()
         ddc::Chunk derivs_mixed_lhs1_lhs2_host_alloc(derivs_domain12, ddc::HostAllocator<double>());
         ddc::ChunkSpan const derivs_mixed_lhs1_lhs2_host
                 = derivs_mixed_lhs1_lhs2_host_alloc.span_view();
-        ddc::Chunk derivs_mixed_lhs1_rhs2_host_alloc(derivs_domain12, ddc::HostAllocator<double>());
-        ddc::ChunkSpan const derivs_mixed_lhs1_rhs2_host
-                = derivs_mixed_lhs1_rhs2_host_alloc.span_view();
         ddc::Chunk derivs_mixed_rhs1_lhs2_host_alloc(derivs_domain12, ddc::HostAllocator<double>());
         ddc::ChunkSpan const derivs_mixed_rhs1_lhs2_host
                 = derivs_mixed_rhs1_lhs2_host_alloc.span_view();
+        ddc::Chunk derivs_mixed_lhs1_rhs2_host_alloc(derivs_domain12, ddc::HostAllocator<double>());
+        ddc::ChunkSpan const derivs_mixed_lhs1_rhs2_host
+                = derivs_mixed_lhs1_rhs2_host_alloc.span_view();
         ddc::Chunk derivs_mixed_rhs1_rhs2_host_alloc(derivs_domain12, ddc::HostAllocator<double>());
         ddc::ChunkSpan const derivs_mixed_rhs1_rhs2_host
                 = derivs_mixed_rhs1_rhs2_host_alloc.span_view();
@@ -515,16 +515,16 @@ void Batched3dSplineTest()
                                                                     deriv_idx1 + shift - 1,
                                                                     deriv_idx2 + shift - 1,
                                                                     0);
-                    derivs_mixed_lhs1_rhs2_host(e) = evaluator
-                                                             .deriv(x0<I1>(),
-                                                                    xN<I2>(),
+                    derivs_mixed_rhs1_lhs2_host(e) = evaluator
+                                                             .deriv(xN<I1>(),
+                                                                    x0<I2>(),
                                                                     x3,
[…]
