### 2025-10-14 a3e4a4d
Rename thr deriv chunkspans to not use trailing underscores
