+++ tests/splines/batched_3d_spline_builder.cpp
@@ -283,7 +283,7 @@ void Batched3dSplineTest()
     int const shift = s_degree % 2; // shift = 0 for even order, 1 for odd order
 
     ddc::Chunk derivs1_lhs_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
-    ddc::ChunkSpan const derivs1_lhs = derivs1_lhs_alloc.span_view();
+    ddc::ChunkSpan const derivs1_lhs_view = derivs1_lhs_alloc.span_view();
     if (s_bcl == ddc::BoundCond::HERMITE) {
         ddc::Chunk derivs1_lhs1_host_alloc(
                 ddc::DiscreteDomain<
@@ -302,19 +302,20 @@ void Batched3dSplineTest()
                             = evaluator.deriv(x0<I1>(), x2, x3, deriv_idx + shift - 1, 0, 0);
                 });
         auto derivs1_lhs1_alloc = ddc::create_mirror_view_and_copy(exec_space, derivs1_lhs1_host);
-        ddc::ChunkSpan const derivs1_lhs1 = derivs1_lhs1_alloc.span_view();
+        ddc::ChunkSpan const derivs1_lhs = derivs1_lhs1_alloc.span_view();
 
         ddc::parallel_for_each(
                 exec_space,
-                derivs1_lhs.domain(),
+                derivs1_lhs_view.domain(),
                 KOKKOS_LAMBDA(
-                        typename decltype(derivs1_lhs.domain())::discrete_element_type const e) {
-                    derivs1_lhs(e) = derivs1_lhs1(DElem<ddc::Deriv<I1>, DDimI2, DDimI3>(e));
+                        typename decltype(derivs1_lhs_view
+                                                  .domain())::discrete_element_type const e) {
+                    derivs1_lhs_view(e) = derivs1_lhs(DElem<ddc::Deriv<I1>, DDimI2, DDimI3>(e));
                 });
     }
 
     ddc::Chunk derivs1_rhs_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
-    ddc::ChunkSpan const derivs1_rhs = derivs1_rhs_alloc.span_view();
+    ddc::ChunkSpan const derivs1_rhs_view = derivs1_rhs_alloc.span_view();
     if (s_bcr == ddc::BoundCond::HERMITE) {
         ddc::Chunk derivs1_rhs1_host_alloc(
                 ddc::DiscreteDomain<
@@ -333,19 +334,20 @@ void Batched3dSplineTest()
                             = evaluator.deriv(xN<I1>(), x2, x3, deriv_idx + shift - 1, 0, 0);
                 });
         auto derivs1_rhs1_alloc = ddc::create_mirror_view_and_copy(exec_space, derivs1_rhs1_host);
-        ddc::ChunkSpan const derivs1_rhs1 = derivs1_rhs1_alloc.span_view();
+        ddc::ChunkSpan const derivs1_rhs = derivs1_rhs1_alloc.span_view();
 
         ddc::parallel_for_each(
                 exec_space,
-                derivs1_rhs.domain(),
+                derivs1_rhs_view.domain(),
                 KOKKOS_LAMBDA(
-                        typename decltype(derivs1_rhs.domain())::discrete_element_type const e) {
[…]
