### 2025-10-14 5da91cc
Add derivs to the characteristics advection example
