+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -193,7 +193,7 @@ public:
     /**
      * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
      * product of 1D Deriv domain and batch domain) in the first dimension, to be passed as
-     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     * argument to the builder, preserving the underlying memory layout (order of dimensions).
      *
      * @tparam The batched discrete domain on which the interpolation points are defined.
      *
@@ -228,7 +228,7 @@ public:
     /**
      * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
      * product of 1D Deriv domain and batch domain) in the second dimension, to be passed as
-     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     * argument to the builder, preserving the underlying memory layout (order of dimensions).
      *
      * @tparam The batched discrete domain on which the interpolation points are defined.
      *
@@ -263,7 +263,7 @@ public:
     /**
      * @brief The type of the whole Derivs domain (1D dimension of interest and cartesian
      * product of 1D Deriv domain and batch domain) in the third dimension, to be passed as
-     * argument to the buider, preserving the underlying memory layout (order of dimensions).
+     * argument to the builder, preserving the underlying memory layout (order of dimensions).
      *
      * @tparam The batched discrete domain on which the interpolation points are defined.
      *
