### 2025-10-14 0bbd1b4
Fix missing doc comment
