### 2025-10-14 1c230b5
Fix some more unused variable warnings
