### 2025-10-14 6404043
Rename some variable to avoid shadowing
