+++ tests/splines/batched_3d_spline_builder.cpp
@@ -282,27 +282,27 @@ void Batched3dSplineTest()
     // Allocate and fill a chunk containing derivs to be passed as input to spline_builder.
     int const shift = s_degree % 2; // shift = 0 for even order, 1 for odd order
 
-    ddc::Chunk derivs1_lhs_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
-    ddc::ChunkSpan const derivs1_lhs_view = derivs1_lhs_alloc.span_view();
+    ddc::Chunk derivs1_lhs_view_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
+    ddc::ChunkSpan const derivs1_lhs_view = derivs1_lhs_view_alloc.span_view();
     if (s_bcl == ddc::BoundCond::HERMITE) {
-        ddc::Chunk derivs1_lhs1_host_alloc(
+        ddc::Chunk derivs1_lhs_host_alloc(
                 ddc::DiscreteDomain<
                         ddc::Deriv<I1>,
                         DDimI2,
                         DDimI3>(derivs_domain1, interpolation_domain2, interpolation_domain3),
                 ddc::HostAllocator<double>());
-        ddc::ChunkSpan const derivs1_lhs1_host = derivs1_lhs1_host_alloc.span_view();
+        ddc::ChunkSpan const derivs1_lhs_host = derivs1_lhs_host_alloc.span_view();
         ddc::for_each(
-                derivs1_lhs1_host.domain(),
+                derivs1_lhs_host.domain(),
                 KOKKOS_LAMBDA(ddc::DiscreteElement<ddc::Deriv<I1>, DDimI2, DDimI3> const e) {
                     auto deriv_idx = ddc::DiscreteElement<ddc::Deriv<I1>>(e).uid();
                     auto x2 = ddc::coordinate(ddc::DiscreteElement<DDimI2>(e));
                     auto x3 = ddc::coordinate(ddc::DiscreteElement<DDimI3>(e));
-                    derivs1_lhs1_host(e)
+                    derivs1_lhs_host(e)
                             = evaluator.deriv(x0<I1>(), x2, x3, deriv_idx + shift - 1, 0, 0);
                 });
-        auto derivs1_lhs1_alloc = ddc::create_mirror_view_and_copy(exec_space, derivs1_lhs1_host);
-        ddc::ChunkSpan const derivs1_lhs = derivs1_lhs1_alloc.span_view();
+        auto derivs1_lhs_alloc = ddc::create_mirror_view_and_copy(exec_space, derivs1_lhs_host);
+        ddc::ChunkSpan const derivs1_lhs = derivs1_lhs_alloc.span_view();
 
         ddc::parallel_for_each(
                 exec_space,
@@ -314,27 +314,27 @@ void Batched3dSplineTest()
                 });
     }
 
-    ddc::Chunk derivs1_rhs_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
-    ddc::ChunkSpan const derivs1_rhs_view = derivs1_rhs_alloc.span_view();
+    ddc::Chunk derivs1_rhs_view_alloc(dom_derivs1, ddc::KokkosAllocator<double, MemorySpace>());
+    ddc::ChunkSpan const derivs1_rhs_view = derivs1_rhs_view_alloc.span_view();
     if (s_bcr == ddc::BoundCond::HERMITE) {
-        ddc::Chunk derivs1_rhs1_host_alloc(
+        ddc::Chunk derivs1_rhs_host_alloc(
                 ddc::DiscreteDomain<
[…]
