### 2025-10-14 5975a3e
Remove another unused variable
