### 2025-10-14 f566763
Update the spline benchmark
