### 2025-10-15 b027037
Dix clang-tidy warnings
