### 2025-10-15 375821a
Dix clang-tidy warnings
