+++ tests/splines/periodic_spline_builder.cpp
@@ -106,7 +106,8 @@ void TestPeriodicSplineBuilderTestIdentity()
             KOKKOS_LAMBDA(DElemX const ix) { yvals(ix) = evaluator(ddc::coordinate(ix)); });
 
     // 6. Create a empty chunk for the derivatives, as they are not used with periodic boundary conditions
-    ddc::ChunkSpan<double const, ddc::StridedDiscreteDomain<DDimX, ddc::Deriv<DimX>>> derivs;
+    ddc::ChunkSpan<double const, ddc::StridedDiscreteDomain<DDimX, ddc::Deriv<DimX>>> const
+            derivs {};
 
     // 7. Finally build the spline by filling `coef`
     spline_builder(coef.span_view(), yvals.span_cview(), derivs);
