### 2025-10-15 a20557d
Fix a copy-paste error in the 1d builder tests
