+++ tests/splines/batched_spline_builder.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) The DDC development team, see COPYRIGHT.md file
+// Copyright (C) The DDC development team, see COPYRIGHT.md filebatched
 //
 // SPDX-License-Identifier: MIT
 
@@ -204,8 +204,9 @@ void BatchedSplineTest()
     ddc::Chunk derivs_alloc(whole_derivs_domain, ddc::KokkosAllocator<double, MemorySpace>());
     ddc::ChunkSpan const derivs = derivs_alloc.span_view();
 
-    ddc::ChunkSpan const derivs_lhs_view = derivs[interpolation_domain.front()];
     if (s_bcl == ddc::BoundCond::HERMITE) {
+        ddc::ChunkSpan const derivs_lhs_view = derivs[interpolation_domain.front()];
+
         ddc::Chunk derivs_lhs_host_alloc(derivs_domain, ddc::HostAllocator<double>());
         ddc::ChunkSpan const derivs_lhs_host = derivs_lhs_host_alloc.span_view();
         for (int ii = 1; ii < derivs_lhs_host.domain().template extent<ddc::Deriv<I>>() + 1; ++ii) {
@@ -222,8 +223,9 @@ void BatchedSplineTest()
                                 e) { derivs_lhs_view(e) = derivs_lhs(DElem<ddc::Deriv<I>>(e)); });
     }
 
-    ddc::ChunkSpan const derivs_rhs_view = derivs[interpolation_domain.front()];
     if (s_bcr == ddc::BoundCond::HERMITE) {
+        ddc::ChunkSpan const derivs_rhs_view = derivs[interpolation_domain.back()];
+
         ddc::Chunk derivs_rhs_host_alloc(derivs_domain, ddc::HostAllocator<double>());
         ddc::ChunkSpan const derivs_rhs_host = derivs_rhs_host_alloc.span_view();
         for (int ii = 1; ii < derivs_rhs_host.domain().template extent<ddc::Deriv<I>>() + 1; ++ii) {
