### 2025-10-17 #9122 @lwshanbd
I've finished my review of ddc, and it works perfectly on my end. Both the manual and Spack installations were successful. I also compiled and executed the example code without any issues.
