+++ include/ddc/kernels/splines/spline_builder_3d.hpp
@@ -12,6 +12,7 @@
 #include <ddc/ddc.hpp>
 
 #include "spline_builder.hpp"
+#include "spline_builder_2d.hpp"
 
 namespace ddc {
 
@@ -52,30 +53,28 @@ public:
     using builder_type1 = ddc::
             SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
 
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate along second dimension.
-    using builder_type2 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate along third dimension.
-    using builder_type3 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower3, BcUpper3, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate the second-dimension-derivatives along first dimension.
-    using builder_deriv_type1 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver>;
-
-    /// @brief The type of the SplineBuilder used by this class to spline-approximate the third-dimension-derivatives along second dimension.
-    using builder_deriv_type2 = ddc::
-            SplineBuilder<ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver>;
+    /// @brief The type of SplineBuilder used by this class to spline-approximate along the second and third dimensions.
+    using builder_type_2_3 = ddc::SplineBuilder2D<
+            ExecSpace,
+            MemorySpace,
+            BSpline2,
+            BSpline3,
+            DDimI2,
+            DDimI3,
+            BcLower2,
+            BcUpper2,
+            BcLower3,
+            BcUpper3,
+            Solver>;
 
     /// @brief The type of the first interpolation continuous dimension.
     using continuous_dimension_type1 = typename builder_type1::continuous_dimension_type;
 
     /// @brief The type of the second interpolation continuous dimension.
-    using continuous_dimension_type2 = typename builder_type2::continuous_dimension_type;
+    using continuous_dimension_type2 = typename builder_type_2_3::continuous_dimension_type1;
 
     /// @brief The type of the third interpolation continuous dimension.
-    using continuous_dimension_type3 = typename builder_type3::continuous_dimension_type;
+    using continuous_dimension_type3 = typename builder_type_2_3::continuous_dimension_type2;
 
     /// @brief The type of the first interpolation discrete dimension.
     using interpolation_discrete_dimension_type1 =
@@ -83,29 +82,29 @@ public:
 
     /// @brief The type of the second interpolation discrete dimension.
     using interpolation_discrete_dimension_type2 =
-            typename builder_type2::interpolation_discrete_dimension_type;
+            typename builder_type_2_3::interpolation_discrete_dimension_type1;
 
     /// @brief The type of the third interpolation discrete dimension.
     using interpolation_discrete_dimension_type3 =
-            typename builder_type3::interpolation_discrete_dimension_type;
+            typename builder_type_2_3::interpolation_discrete_dimension_type2;
 
     /// @brief The type of the B-splines in the first dimension.
     using bsplines_type1 = typename builder_type1::bsplines_type;
 
     /// @brief The type of the B-splines in the second dimension.
-    using bsplines_type2 = typename builder_type2::bsplines_type;
+    using bsplines_type2 = typename builder_type_2_3::bsplines_type1;
 
[…]
