+++ search/all_12.js
@@ -5,7 +5,7 @@ var searchData=
   ['only_2',['Quick Start (core features only)',['../installation.html#autotoc_md32',1,'']]],
   ['operator_20bool_3',['operator bool',['../classddc_1_1DiscreteDomain_3_4.html#a6c35d71dd28805fa4b1c99d9db306297',1,'ddc::DiscreteDomain&lt;&gt;::operator bool()'],['../classddc_1_1StridedDiscreteDomain_3_4.html#aa454559e81b5d820e09abef53edaaf5b',1,'ddc::StridedDiscreteDomain&lt;&gt;::operator bool()'],['../classddc_1_1StridedDiscreteDomain.html#a128bd793b520e9ecf74d727df4060c49',1,'ddc::StridedDiscreteDomain::operator bool()'],['../classddc_1_1DiscreteDomain.html#af072b07bd07fce19032545ad54104876',1,'ddc::DiscreteDomain::operator bool()'],['../classddc_1_1SparseDiscreteDomain.html#ab3abcbde1bd52193157bec9d23ebab36',1,'ddc::SparseDiscreteDomain::operator bool()'],['../classddc_1_1SparseDiscreteDomain_3_4.html#ab2a48b86b7919fc41e8d58eef0380202',1,'ddc::SparseDiscreteDomain&lt;&gt;::operator bool()']]],
   ['operator_21_3d_4',['operator!=',['../classddc_1_1DiscreteDomain.html#a5468b388ccf9e74437938c60f8aca7dd',1,'ddc::DiscreteDomain::operator!=()'],['../namespaceddc.html#ae4269f314ca30fb3ded4fdcc3a399965',1,'ddc::operator!=(KokkosAllocator&lt; T, MST &gt; const &amp;, KokkosAllocator&lt; U, MSU &gt; const &amp;) noexcept'],['../namespaceddc.html#af71143b67a010f2229acda8c9a66e21e',1,'ddc::operator!=(DiscreteElement&lt; Tags... &gt; const &amp;lhs, DiscreteElement&lt; OTags... &gt; const &amp;rhs) noexcept'],['../namespaceddc.html#a1e2daade1cf2aa66fd6772beb01ac888',1,'ddc::operator!=(AlignedAllocator&lt; T, NT &gt; const &amp;, AlignedAllocator&lt; U, NU &gt; const &amp;) noexcept'],['../structddc_1_1StridedDiscreteDomainIterator.html#ae8138543e874a24b2b3268fff4badf93',1,'ddc::StridedDiscreteDomainIterator::operator!='],['../classddc_1_1StridedDiscreteDomain_3_4.html#a089570efbe37329cab278da7d0b928c8',1,'ddc::StridedDiscreteDomain&lt;&gt;::operator!=()'],['../classddc_1_1DiscreteDomain_3_4.html#a5678f3ad8565423bbb00e169ec9b1415',1,'ddc::DiscreteDomain&lt;&gt;::operator!=()'],['../structddc_1_1DiscreteDomainIterator.html#a6f6e25808499026840dbb113394b7505',1,'ddc::DiscreteDomainIterator::operator!='],['../classddc_1_1DiscreteVector.html#a62e896bfa78045521042a0eab87c0acb',1,'ddc::DiscreteVector::operator!=()'],['../classddc_1_1SparseDiscreteDomain.html#ac9a4775dd525df009e13773729edc4c0',1,'ddc::SparseDiscreteDomain::operator!=()'],['../classddc_1_1SparseDiscreteDomain_3_4.html#a9968f454afa63adf5a4428605a1c7de7',1,'ddc::SparseDiscreteDomain&lt;&gt;::operator!=()'],['../classddc_1_1StridedDiscreteDomain.html#a8ba64e3d66da3190e94722d984
[…]
