+++ hierarchy.js
@@ -90,10 +90,12 @@ var hierarchy =
     [ "ddc::SparseDiscreteDomainIterator< DDim >", "namespaceddc.html#structddc_1_1SparseDiscreteDomainIterator", null ],
     [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSplines, InterpolationDDim, BcLower, BcUpper, Solver >", "classddc_1_1SplineBuilder.html", null ],
     [ "ddc::SplineBuilder2D< ExecSpace, MemorySpace, BSpline1, BSpline2, DDimI1, DDimI2, BcLower1, BcUpper1, BcLower2, BcUpper2, Solver >", "classddc_1_1SplineBuilder2D.html", null ],
+    [ "ddc::SplineBuilder2D< ExecSpace, MemorySpace, BSpline2, BSpline3, DDimI2, DDimI3, BcLower2, BcUpper2, BcLower3, BcUpper3, Solver >", "classddc_1_1SplineBuilder2D.html", null ],
     [ "ddc::SplineBuilder3D< ExecSpace, MemorySpace, BSpline1, BSpline2, BSpline3, DDimI1, DDimI2, DDimI3, BcLower1, BcUpper1, BcLower2, BcUpper2, BcLower3, BcUpper3, Solver >", "classddc_1_1SplineBuilder3D.html", null ],
     [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSpline1, DDimI1, BcLower1, BcUpper1, Solver >", "classddc_1_1SplineBuilder.html", null ],
+    [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower1, BcUpper1, Solver >", "classddc_1_1SplineBuilder.html", null ],
     [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSpline2, DDimI2, BcLower2, BcUpper2, Solver >", "classddc_1_1SplineBuilder.html", null ],
-    [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower3, BcUpper3, Solver >", "classddc_1_1SplineBuilder.html", null ],
+    [ "ddc::SplineBuilder< ExecSpace, MemorySpace, BSpline3, DDimI3, BcLower2, BcUpper2, Solver >", "classddc_1_1SplineBuilder.html", null ],
     [ "ddc::SplineEvaluator< ExecSpace, MemorySpace, BSplines, EvaluationDDim, LowerExtrapolationRule, UpperExtrapolationRule >", "classddc_1_1SplineEvaluator.html", null ],
     [ "ddc::SplineEvaluator2D< ExecSpace, MemorySpace, BSplines1, BSplines2, EvaluationDDim1, EvaluationDDim2, LowerExtrapolationRule1, UpperExtrapolationRule1, LowerExtrapolationRule2, UpperExtrapolationRule2 >", "classddc_1_1SplineEvaluator2D.html", null ],
     [ "ddc::SplineEvaluator3D< ExecSpace, MemorySpace, BSplines1, BSplines2, BSplines3, EvaluationDDim1, EvaluationDDim2, EvaluationDDim3, LowerExtrapolationRule1, UpperExtrapolationRule1, LowerExtrapolationRule2, UpperExtrapolationRule2, LowerExtrapolationRule3, UpperExtrapolationRule3 >", "classddc_1_1SplineEvaluator3D.html", null ],
