+++ search/typedefs_2.js
@@ -10,8 +10,8 @@ var searchData=
   ['const_5fallocation_5fmdspan_5ftype_7',['const_allocation_mdspan_type',['../classddc_1_1ChunkCommon.html#a2c32334f535229d4b0324492cd77b188',1,'ddc::ChunkCommon::const_allocation_mdspan_type'],['../classddc_1_1Chunk.html#adfe7a4cd89342fb16202e5f49b42e83d',1,'ddc::Chunk::const_allocation_mdspan_type'],['../classddc_1_1ChunkSpan.html#aabf9e3daf94bab0e7db2848f05012fb9',1,'ddc::ChunkSpan::const_allocation_mdspan_type']]],
   ['continuous_5fdimension_5ftype_8',['continuous_dimension_type',['../classddc_1_1UniformPointSampling.html#a9b60e05af4dc496085f30d785d4d66be',1,'ddc::UniformPointSampling::continuous_dimension_type'],['../classddc_1_1PeriodicSampling.html#ab7065f0210b99346a966d4353ae25512',1,'ddc::PeriodicSampling::continuous_dimension_type'],['../classddc_1_1NonUniformPointSampling.html#a48388f252b0320a2a8e02d82581dfb9f',1,'ddc::NonUniformPointSampling::continuous_dimension_type'],['../classddc_1_1SplineEvaluator.html#aa7e0c79516dff00341775124dc69d61d',1,'ddc::SplineEvaluator::continuous_dimension_type'],['../classddc_1_1SplineBuilder.html#a872b6c87ce8ca70b8456e18d97bd2ac0',1,'ddc::SplineBuilder::continuous_dimension_type'],['../classddc_1_1UniformBSplines.html#a150fd07672395d35efaeae1d40acd1df',1,'ddc::UniformBSplines::continuous_dimension_type'],['../classddc_1_1NonUniformBSplines.html#a228f5629c111ccbdfc480e13ea517cb2',1,'ddc::NonUniformBSplines::continuous_dimension_type']]],
   ['continuous_5fdimension_5ftype1_9',['continuous_dimension_type1',['../classddc_1_1SplineBuilder2D.html#a07d5c490a3cbfd9d3070f63e99d997d0',1,'ddc::SplineBuilder2D::continuous_dimension_type1'],['../classddc_1_1SplineBuilder3D.html#aff3ff40ee5ec290a01c7295c336278f4',1,'ddc::SplineBuilder3D::continuous_dimension_type1'],['../classddc_1_1SplineEvaluator2D.html#aa875b246a469a139fb191088e3e22cb8',1,'ddc::SplineEvaluator2D::continuous_dimension_type1'],['../classddc_1_1SplineEvaluator3D.html#a881b156d23515b5d4db085971cd9bc48',1,'ddc::SplineEvaluator3D::continuous_dimension_type1']]],
[…]

[…]
