### 2025-10-31 a5b732e
Update more citations
