### 2025-10-31 #9122 @tpadioleau
> I performed all the tests and examples (2 DDC::core and the DDC::fft) without issues. I only have some minor remarks:
> - (ddc/docs/installation.md) The Quick Start does not specify that CMake should be executed within the DDC folder. I suggest adding something like cd ddc in the bash console instructions between the git command the first CMake.

Actually it was a typo in the command line instruction. It has been fixed in https://github.com/CExA-project/ddc/pull/957.
