### 2025-10-31 162c8e0
Update citations for Gyselalib++ and kokkos-fft. Add Baptiste as author (#956)
* Update references

* Update authors and date

* Update more citations

---------

Co-authored-by: Thomas Padioleau <thomas.padioleau@cea.fr>
