### 2025-10-31 113f78b
minor changes for JOSS publication (#955)
* Update paper.md

* Update paper.bib

---------

Co-authored-by: Thomas Padioleau <thomas.padioleau@cea.fr>
