+++ paper/paper.md
@@ -1,35 +1,25 @@
-# DDC: The Discrete Domain Computation library
-
 ---
 title: 'DDC: The Discrete Domain Computation library'
 tags:
-
 - C++
 - HPC
 - labelled arrays
 - xarray
-
 authors:
-
 - name: Thomas Padioleau
   orcid: 0000-0001-5496-0013
   equal-contrib: true
   affiliation: 1
-
 - name: Julien Bigot
   orcid: 0000-0002-0015-4304
   affiliation: 1
-
 - name: Emily Bourne
   orcid: 0000-0002-3469-2338
   affiliation: 2
-
 - name: Baptiste Legouix
   orcid: 0009-0006-7585-669X
   affiliation: 3
-
 affiliations:
-
 - name: Université Paris-Saclay, UVSQ, CNRS, CEA, Maison de la Simulation, 91191, Gif-sur-Yvette, France
   index: 1
 - name: SCITAS, EPFL, CH-1015 Lausanne, Switzerland
@@ -38,7 +28,6 @@ affiliations:
   index: 3
 date: 31 October 2025
 bibliography: paper.bib
-
 ---
 
 ## Summary
@@ -51,11 +40,11 @@ The use of multidimensional arrays is widespread across various fields, particul
 
 Many programming languages commonly used in scientific computing support multidimensional arrays in different ways. Fortran, a longstanding choice in the field, and Julia, a more recent language, both natively support these data structures. In contrast, the Python ecosystem relies on the popular NumPy library’s `numpy.Array` [@harris2020array]. Meanwhile, C++23 introduced `std::mdspan` to the standard library. This container was inspired by `Kokkos::View` from the Kokkos library which also serves as the foundation of DDC.
 
-Despite their importance, multidimensional arrays introduce several practical challenges. In a sense, they encourage the usage of implicit information in the source code. A frequent source of errors is the inadvertent swapping of indices when accessing elements. Such errors can be difficult to detect, especially given the common convention of using single-letter variable names like `i` and `j` for indexing. Another challenge in medium to large codebases is the lack of semantic clarity in function signatures when using raw multidimensional arrays. When array dimensions carry specific meanings, this information is not explicitly represented in the source code, leaving it up to the user to ensure that dimensions are ordered correctly according to implicit expectations. For example it is quite usual to use the same index for multiple interpretations: looping over mesh cells identified by `i` and interpreting `i+1` as the face to the right. Another example is slicing that removes dimensions, this can shift the positions of remaining dimensions, altering the correspondence between axis indices and their semantic meanings.
+Despite their importance, multidimensional arrays introduce several practical challenges. In a sense, they encourage the usage of implicit information in the source code. A frequent source of errors is the inadvertent swapping of indices when accessing elements. Such errors can be difficult to detect, especially given the common convention of using single-letter variable names like `i` and `j` for indexing. Another challenge in medium-to-large codebases is the lack of semantic clarity in function signatures when using raw multidimensional arrays. When array dimensions carry specific meanings, this information is not explicitly represented in the source code, leaving it up to the user to ensure that dimensions are ordered correctly according to implicit expectations. For example it is quite usual to use the same index for multiple interpretations: looping over mesh cells identified by `i` and interpreting `i+1` as the face to the right. Another example is slicing that removes dimensions, this can shift the positions of remaining dimensions, altering the correspondence between axis indices and their semantic meanings.
 
[…]
