+++ paper/paper.bib
@@ -73,7 +73,7 @@ SPDX-License-Identifier: MIT
 }
 
 @article{grandgirard2016,
-title = {A 5D gyrokinetic full-f global semi-Lagrangian code for flux-driven ion turbulence simulations},
+title = {A 5D gyrokinetic full-f global semi-{L}agrangian code for flux-driven ion turbulence simulations},
 journal = {Computer Physics Communications},
 volume = {207},
 pages = {35-68},
