### 2025-10-31 83953ea
Warn about memory issues during installation with Spack (#958)
