### 2025-10-31 bd6671b
Update paper.bib
