+++ paper/paper.bib
@@ -111,7 +111,7 @@ author = {V. Grandgirard and J. Abiteboul and J. Bigot and T. Cartier-Michaud an
     number = {113},
     pages = {8582},
     author = {Bourne, Emily and Grandgirard, Virginie and Asahi, Yuuichi and Bigot, Julien and Donnel, Peter and Hoffmann, Alexander and Kara, Abdelhadi and Krah, Philipp and Legouix, Baptiste and Malaboeuf, Etienne and Munschy, Yann and Obrejan, Kevin and Padioleau, Thomas and Protais, Matthieu and Vidal, Pauline},
-    title = {Gyselalib++: A Portable C++ Library for Semi-Lagrangian Kinetic and Gyrokinetic Simulations},
+    title = {Gyselalib++: A Portable C++ Library for Semi-{L}agrangian Kinetic and Gyrokinetic Simulations},
     journal = {Journal of Open Source Software}
 }
 
