### 2025-10-31 #9122 @danielskatz
> [@danielskatz](https://github.com/danielskatz) Thank you for your help, regarding
>
> > Make a tagged release of your software, and list the version tag of the archived version here.
>
> Can we use the existing release `v0.8.0` (from June 5) that reviewers likely used to install with Spack ? Some developments were merged a few days ago that were likely not part of the review.

That should be ok if there weren't changes to the software in the review process. Ideally, the archived version should be what was signed off by the reviewers at the end of the review process.
