### 2025-10-31 f7be2d7
Rename some variables
