+++ .github/workflows/tests-macos.yaml
@@ -199,7 +199,7 @@ jobs:
       - name: Run unit tests
         run: ctest --test-dir build --output-on-failure --timeout 10 --output-junit tests.xml
       - name: Publish Test Report
-        uses: mikepenz/action-junit-report@5b7ee5a21e8674b695313d769f3cbdfd5d4d53a4  # v6.0.0
+        uses: mikepenz/action-junit-report@e08919a3b1fb83a78393dfb775a9c37f17d8eea6  # v6.0.1
         if: ( success() || failure() )  # always run even if the previous step fails
         with:
           annotate_only: true  # forked repo cannot write to checks so just do annotations
