+++ CMakeLists.txt
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: MIT
 
 cmake_minimum_required(VERSION 3.25)
-project(DDC VERSION 0.8.0 LANGUAGES CXX)
+project(DDC VERSION 0.9.0 LANGUAGES CXX)
 
 # List of options
 
