### 2025-11-03 8c7f2a1
Bump mikepenz/action-junit-report from 6.0.0 to 6.0.1 (#962)
Bumps [mikepenz/action-junit-report](https://github.com/mikepenz/action-junit-report) from 6.0.0 to 6.0.1.
- [Release notes](https://github.com/mikepenz/action-junit-report/releases)
- [Commits](https://github.com/mikepenz/action-junit-report/compare/5b7ee5a21e8674b695313d769f3cbdfd5d4d53a4...e08919a3b1fb83a78393dfb775a9c37f17d8eea6)

---
updated-dependencies:
- dependency-name: mikepenz/action-junit-report
  dependency-version: 6.0.1
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: 
[…]
