### 2025-11-03 #9122 @tpadioleau
As you requested:
- I have created the release [v0.9.0](https://github.com/CExA-project/ddc/releases/tag/v0.9.0)
- it is now archived on zenodo with DOI https://doi.org/10.5281/zenodo.17511922
