### 2025-11-05 77e40a0
Add ND evaluator - works for host backends for the moment
