+++ include/ddc/kernels/splines/spline_evaluator_nd.hpp
@@ -1,0 +1,975 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include <array>
+#include <cstddef>
+#include <tuple>
+#include <type_traits>
+#include <utility>
+
+#include <ddc/ddc.hpp>
+
+#include <Kokkos_Core.hpp>
+
+#include "ddc/coordinate.hpp"
+#include "ddc/detail/type_seq.hpp"
+#include "ddc/discrete_domain.hpp"
+
+#include "integrals.hpp"
+#include "periodic_extrapolation_rule.hpp"
+
+namespace ddc {
+
+namespace detail {
+template <typename... Args>
+class SplineEvaluatorND;
+
+template <
+        typename ExecSpace,
+        typename MemorySpace,
+        typename... BSplines,
+        typename... EvaluationDDim,
+        typename... LowerExtrapolationRule,
+        typename... UpperExtrapolationRule,
+        std::size_t... Idx>
+class SplineEvaluatorND<
+        ExecSpace,
+        MemorySpace,
+        ddc::detail::TypeSeq<BSplines...>,
+        ddc::detail::TypeSeq<EvaluationDDim...>,
+        ddc::detail::TypeSeq<LowerExtrapolationRule...>,
+        ddc::detail::TypeSeq<UpperExtrapolationRule...>,
+        std::index_sequence<Idx...>>
+{
+private:
+    /**
+     * @brief Tag to indicate that the value of the spline should be evaluated.
+     */
+    struct eval_type
+    {
+    };
+
+    /**
+     * @brief Tag to indicate that derivative of the spline should be evaluated.
+     */
+    struct eval_deriv_type
+    {
+    };
+
+    using bsplines_ts = ddc::detail::TypeSeq<BSplines...>;
+    using evaluation_ddim_ts = ddc::detail::TypeSeq<EvaluationDDim...>;
+    using lower_extrap_rule_ts = ddc::detail::TypeSeq<LowerExtrapolationRule...>;
+    using upper_extrap_rule_ts = ddc::detail::TypeSeq<UpperExtrapolationRule...>;
+
+    static constexpr std::size_t dimension = sizeof...(Idx);
+
+    template <std::size_t I, std::size_t... Others>
+    static constexpr bool index_seq_contains()
+    {
+        return ((I == Others) || ...);
+    }
+
+public:
+    /// @brief The type of the Ith evaluation continuous dimension used by this class.
+    /// @tparam I the requested dimension
+    template <std::size_t I>
+    using continuous_dimension_type =
+            typename ddc::type_seq_element_t<I, bsplines_ts>::continuous_dimension_type;
+
+    /// @brief The type of the Kokkos execution space used by this class.
+    using exec_space = ExecSpace;
+
+    /// @brief The type of the Kokkos memory space used by this class.
+    using memory_space = MemorySpace;
+
+    /// @brief The type of the Ith discrete dimension of interest used by this class.
+    template <std::size_t I>
+    using evaluation_discrete_dimension_type = ddc::type_seq_element_t<I, evaluation_ddim_ts>;
+
+    /// @brief The discrete dimension representing the B-splines along Ith dimension.
+    template <std::size_t I>
+    using bsplines_type = ddc::type_seq_element_t<I, bsplines_ts>;
+
+    /**
+     * @brief The type of the domain for the 1D, 2D, ... or ND evaluation mesh along specified dimensions used by this class.
+     *
+     * @tparam Dims the required dimensions, 0 indexed
+     */
+    template <std::size_t... Dims>
+    using evaluation_domain_type = ddc::DiscreteDomain<evaluation_discrete_dimension_type<Dims>...>;
+
[…]
