### 2025-11-05 ea37228
Add ND evaluator - works for host backends for the moment
