+++ tests/splines/batched_nd_spline_builder.cpp
@@ -1,0 +1,1423 @@
+// Copyright (C) The DDC development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT
+
+#include <algorithm>
+#include <cstddef>
+#if defined(BC_HERMITE)
+#    include <optional>
+#endif
+#if defined(BSPLINES_TYPE_UNIFORM)
+#    include <type_traits>
+#endif
+#include <vector>
+
+#include <ddc/ddc.hpp>
+#include <ddc/kernels/splines.hpp>
+
+#include <gtest/gtest.h>
+
+#include <Kokkos_Core.hpp>
+
+#include "evaluator_3d.hpp"
+#if defined(BC_PERIODIC)
+#    include "cosine_evaluator.hpp"
+#else
+#    include "polynomial_evaluator.hpp"
+#endif
+#include "spline_error_bounds.hpp"
+
+inline namespace anonymous_namespace_workaround_batched_3d_spline_builder_cpp {
+
+#if defined(BC_PERIODIC)
+struct DimX
+{
+    static constexpr bool PERIODIC = true;
+};
+
+struct DimY
+{
+    static constexpr bool PERIODIC = true;
+};
+
+struct DimZ
+{
+    static constexpr bool PERIODIC = true;
+};
+#else
+
+struct DimX
+{
+    static constexpr bool PERIODIC = false;
+};
+
+struct DimY
+{
+    static constexpr bool PERIODIC = false;
+};
+
+struct DimZ
+{
+    static constexpr bool PERIODIC = false;
+};
+#endif
+
+struct DDimBatch
+{
+};
+
+constexpr std::size_t s_degree = DEGREE;
+
+#if defined(BC_PERIODIC)
+constexpr ddc::BoundCond s_bcl = ddc::BoundCond::PERIODIC;
+constexpr ddc::BoundCond s_bcr = ddc::BoundCond::PERIODIC;
+#elif defined(BC_GREVILLE)
+constexpr ddc::BoundCond s_bcl = ddc::BoundCond::GREVILLE;
+constexpr ddc::BoundCond s_bcr = ddc::BoundCond::GREVILLE;
+#elif defined(BC_HERMITE)
+constexpr ddc::BoundCond s_bcl = ddc::BoundCond::HERMITE;
+constexpr ddc::BoundCond s_bcr = ddc::BoundCond::HERMITE;
+#endif
+
+template <typename BSpX>
+using GrevillePoints = ddc::GrevilleInterpolationPoints<BSpX, s_bcl, s_bcr>;
+
+#if defined(BSPLINES_TYPE_UNIFORM)
+template <typename X>
+struct BSplines : ddc::UniformBSplines<X, s_degree>
+{
+};
+#elif defined(BSPLINES_TYPE_NON_UNIFORM)
+template <typename X>
+struct BSplines : ddc::NonUniformBSplines<X, s_degree>
+{
+};
+#endif
+
+// In the dimensions of interest, the discrete dimension is deduced from the Greville points type.
+template <typename X>
+struct DDimGPS : GrevillePoints<BSplines<X>>::interpolation_discrete_dimension_type
+{
+};
+
+#if defined(BC_PERIODIC)
+template <typename DDim1, typename DDim2, typename DDim3>
+using evaluator_type = Evaluator3D::Evaluator<
+        CosineEvaluator::Evaluator<DDim1>,
+        CosineEvaluator::Evaluator<DDim2>,
+        CosineEvaluator::Evaluator<DDim3>>;
+#else
+template <typename DDim1, typename DDim2, typename DDim3>
+using evaluator_type = Evaluator3D::Evaluator<
[…]
