### 2025-11-06 #9122 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
