### 2025-11-07 4a3eed2
Remove a variadic template argument in one of the `deriv` overloads
