+++ tests/splines/batched_nd_spline_builder.cpp
@@ -27,7 +27,7 @@
 #endif
 #include "spline_error_bounds.hpp"
 
-inline namespace anonymous_namespace_workaround_batched_3d_spline_builder_cpp {
+inline namespace anonymous_namespace_workaround_batched_nd_spline_builder_cpp {
 
 #if defined(BC_PERIODIC)
 struct DimX
@@ -174,7 +174,7 @@ template <
         typename DDimI2,
         typename DDimI3,
         typename... DDims>
-void Batched3dSplineTest()
+void BatchedNdSplineTest()
 {
     using I1 = typename DDimI1::continuous_dimension_type;
     using I2 = typename DDimI2::continuous_dimension_type;
@@ -1117,25 +1117,25 @@ void Batched3dSplineTest()
 
     // Call spline_evaluator on the same mesh we started with
     spline_evaluator(spline_eval, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator
-            .template deriv<I1>(spline_eval_deriv1, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator
-            .template deriv<I2>(spline_eval_deriv2, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator
-            .template deriv<I3>(spline_eval_deriv3, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator.template deriv<
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
+            I1>>(spline_eval_deriv1, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
+            I2>>(spline_eval_deriv2, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
+            I3>>(spline_eval_deriv3, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
             I1,
-            I2>(spline_eval_deriv12, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator.template deriv<
+            I2>>(spline_eval_deriv12, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
             I2,
-            I3>(spline_eval_deriv23, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator.template deriv<
+            I3>>(spline_eval_deriv23, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
             I1,
-            I3>(spline_eval_deriv13, coords_eval.span_cview(), coef.span_cview());
-    spline_evaluator.template deriv<
+            I3>>(spline_eval_deriv13, coords_eval.span_cview(), coef.span_cview());
+    spline_evaluator.template deriv<ddc::detail::TypeSeq<
             I1,
             I2,
-            I3>(spline_eval_deriv123, coords_eval.span_cview(), coef.span_cview());
+            I3>>(spline_eval_deriv123, coords_eval.span_cview(), coef.span_cview());
 
     // Checking errors (we recover the initial values)
[…]
