### 2025-11-07 b6e8b14
Remove a variadic template argument in one of the `deriv` overloads
