+++ tests/splines/batched_nd_spline_builder.cpp
@@ -208,17 +208,14 @@ void BatchedNdSplineTest()
             interpolation_domain2,
             interpolation_domain3);
 
+#if defined(BC_HERMITE)
     // Create the derivs domain
-    int const nbc_i1 = s_degree / 2;
-    int const nbc_i2 = s_degree / 2;
-    int const nbc_i3 = s_degree / 2;
-
     ddc::DiscreteDomain<ddc::Deriv<I1>> const
-            derivs_domain1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(nbc_i1));
+            derivs_domain1(DElem<ddc::Deriv<I1>>(1), DVect<ddc::Deriv<I1>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I2>> const
-            derivs_domain2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(nbc_i2));
+            derivs_domain2(DElem<ddc::Deriv<I2>>(1), DVect<ddc::Deriv<I2>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I3>> const
-            derivs_domain3(DElem<ddc::Deriv<I3>>(1), DVect<ddc::Deriv<I3>>(nbc_i3));
+            derivs_domain3(DElem<ddc::Deriv<I3>>(1), DVect<ddc::Deriv<I3>>(s_degree / 2));
     ddc::DiscreteDomain<ddc::Deriv<I1>, ddc::Deriv<I2>, DDimI3> const
             derivs_domain12(derivs_domain1, derivs_domain2, interpolation_domain3);
     ddc::DiscreteDomain<DDimI1, ddc::Deriv<I2>, ddc::Deriv<I3>> const
@@ -236,38 +233,10 @@ void BatchedNdSplineTest()
     auto const dom_derivs23
             = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs2, derivs_domain3);
     auto const dom_derivs13
-            = ddc::replace_dim_of<DDimI1, ddc::Deriv<I1>>(dom_derivs3, derivs_domain1);
+            = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs1, derivs_domain3);
     auto const dom_derivs_all
             = ddc::replace_dim_of<DDimI3, ddc::Deriv<I3>>(dom_derivs12, derivs_domain3);
-
-    auto const whole_derivs_domain1 = ddc::detail::get_whole_derivs_domain<
-            ddc::Deriv<I1>>(interpolation_domain1, dom_vals, nbc_i1);
-    auto const whole_derivs_domain2 = ddc::detail::get_whole_derivs_domain<
-            ddc::Deriv<I2>>(interpolation_domain2, dom_vals, nbc_i2);
-    auto const whole_derivs_domain3 = ddc::detail::get_whole_derivs_domain<
-            ddc::Deriv<I3>>(interpolation_domain3, dom_vals, nbc_i3);
-    auto const whole_derivs_domain1_2
-            = ddc::detail::get_whole_derivs_domain<ddc::Deriv<I1>, ddc::Deriv<I2>>(
-                    ddc::select<DDimI1, DDimI2>(interpolation_domain),
-                    dom_vals,
-                    nbc_i1,
-                    nbc_i2);
-    auto const whole_derivs_domain2_3
-            = ddc::detail::get_whole_derivs_domain<ddc::Deriv<I2>, ddc::Deriv<I3>>(
-                    ddc::select<DDimI2, DDimI3>(interpolation_domain),
-                    dom_vals,
-                    nbc_i2,
[…]
