### 2025-11-07 a953f0f
Add the link to the JOSS doc in the README
