### 2025-11-07 c5bd926
Add the CITATION.cff file
