### 2025-11-07 bd72a4c
Add copyright information
