### 2025-11-07 d8d2395
Remove most of the variadic template parameters - test the 1d ND evaluator - rename the 3d test file
