+++ src/ddc/kernels/splines/spline_evaluator_nd.hpp
@@ -30,18 +30,16 @@ class SplineEvaluatorND;
 template <
         typename ExecSpace,
         typename MemorySpace,
-        typename... BSplines,
-        typename... EvaluationDDim,
-        typename... LowerExtrapolationRule,
-        typename... UpperExtrapolationRule,
+        typename BSplines,
+        typename EvaluationDDim,
+        typename ExtrapolationRules,
         std::size_t... Idx>
 class SplineEvaluatorND<
         ExecSpace,
         MemorySpace,
-        ddc::detail::TypeSeq<BSplines...>,
-        ddc::detail::TypeSeq<EvaluationDDim...>,
-        ddc::detail::TypeSeq<LowerExtrapolationRule...>,
-        ddc::detail::TypeSeq<UpperExtrapolationRule...>,
+        BSplines,
+        EvaluationDDim,
+        ExtrapolationRules,
         std::index_sequence<Idx...>>
 {
 private:
@@ -59,10 +57,12 @@ private:
     {
     };
 
-    using bsplines_ts = ddc::detail::TypeSeq<BSplines...>;
-    using evaluation_ddim_ts = ddc::detail::TypeSeq<EvaluationDDim...>;
-    using lower_extrap_rule_ts = ddc::detail::TypeSeq<LowerExtrapolationRule...>;
-    using upper_extrap_rule_ts = ddc::detail::TypeSeq<UpperExtrapolationRule...>;
+    using bsplines_ts = BSplines;
+    using evaluation_ddim_ts = EvaluationDDim;
+    using lower_extrap_rule_ts
+            = ddc::detail::TypeSeq<ddc::type_seq_element_t<2 * Idx, ExtrapolationRules>...>;
+    using upper_extrap_rule_ts
+            = ddc::detail::TypeSeq<ddc::type_seq_element_t<2 * Idx + 1, ExtrapolationRules>...>;
 
     static constexpr std::size_t dimension = sizeof...(Idx);
 
@@ -128,8 +128,8 @@ public:
     template <
             class BatchedInterpolationDDom,
             class = std::enable_if_t<ddc::is_discrete_domain_v<BatchedInterpolationDDom>>>
-    using batch_domain_type =
-            typename ddc::remove_dims_of_t<BatchedInterpolationDDom, EvaluationDDim...>;
+    using batch_domain_type = ddc::detail::convert_type_seq_to_discrete_domain_t<
+            ddc::type_seq_remove_t<ddc::to_type_seq_t<BatchedInterpolationDDom>, EvaluationDDim>>;
 
     /**
      * @brief The type of the whole spline domain (cartesian product of N:wspline domain
@@ -155,30 +155,30 @@ public:
     using upper_extrapolation_rule_type = ddc::type_seq_element_t<I, upper_extrap_rule_ts>;
 
 private:
-    std::tuple<LowerExtrapolationRule...> m_lower_extrap_rules;
-    std::tuple<UpperExtrapolationRule...> m_upper_extrap_rules;
+    std::tuple<ddc::type_seq_element_t<2 * Idx, ExtrapolationRules>...> m_lower_extrap_rules;
+    std::tuple<ddc::type_seq_element_t<2 * Idx + 1, ExtrapolationRules>...> m_upper_extrap_rules;
 
 public:
     static_assert(
-            sizeof...(BSplines) == sizeof...(Idx),
+            ddc::type_seq_size_v<BSplines> == dimension,
             "Number of BSpline dims should be equal to the dimension");
     static_assert(
-            sizeof...(EvaluationDDim) == sizeof...(Idx),
+            ddc::type_seq_size_v<EvaluationDDim> == dimension,
             "Number of evaluation dims should be equal to the dimensions");
     static_assert(
[…]
