### 2025-11-07 728d816
Remove most of the variadic template parameters - test the 1d ND evaluator - rename the 3d test file
